% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BasisExpansion}
\alias{BasisExpansion}
\title{Returns the weights for the smoothed observed functions}
\arguments{
\item{y_obs}{Field of Vectors containing the observed values}

\item{B_obs}{Field of Matrices containing the basis functions evaluated at the observed time points}

\item{n_funct}{Int containing the number of observed functions}

\item{K}{Int containing the number of clusters}

\item{P}{Int containing the number of Basis Functions}
}
\description{
Returns the weights for the smoothed observed functions
}
