\name{BayesFPMM-package}
\alias{BayesFPMM-package}
\alias{BayesFPMM}
\docType{package}
\title{
\packageTitle{BayesFPMM}
}
\description{
\packageDescription{BayesFPMM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BayesFPMM}
\packageIndices{BayesFPMM}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{BayesFPMM}

Maintainer: \packageMaintainer{BayesFPMM}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
}

