% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BFMMM_Theta_est_Cov_Adj}
\alias{BFMMM_Theta_est_Cov_Adj}
\title{Find initial starting points for parameters given nu, eta, and Z parameters for functional data}
\usage{
BFMMM_Theta_est_Cov_Adj(
  tot_mcmc_iters,
  n_try,
  k,
  Y,
  time,
  X,
  n_funct,
  basis_degree,
  n_eigen,
  boundary_knots,
  internal_knots,
  Z_samp,
  nu_samp,
  eta_samp,
  burnin_prop = 0.8,
  c = NULL,
  b = 10,
  nu_1 = 3,
  alpha1l = 2,
  alpha2l = 3,
  beta1l = 2,
  beta2l = 2,
  a_Z_PM = 10000,
  a_pi_PM = 1000,
  var_alpha3 = 0.05,
  var_epsilon1 = 1,
  var_epsilon2 = 1,
  alpha_nu = 1,
  beta_nu = 10,
  alpha_eta = 1,
  beta_eta = 10,
  alpha_0 = 1,
  beta_0 = 1,
  covariance_adj = FALSE
)
}
\arguments{
\item{tot_mcmc_iters}{Int containing the total number of MCMC iterations}

\item{n_try}{Int containing how many different chains are tried}

\item{k}{Int containing the number of clusters}

\item{Y}{List of vectors containing the observed values}

\item{time}{List of vectors containing the observed time points}

\item{X}{Matrix of covariates (each row corresponds to an observation)}

\item{n_funct}{Int containing the number of functions}

\item{basis_degree}{Int containing the degree of B-splines used}

\item{n_eigen}{Int containing the number of eigenfunctions}

\item{boundary_knots}{Vector containing the boundary points of our index domain of interest}

\item{internal_knots}{Vector location of internal knots for B-splines}

\item{Z_samp}{Cube containing initial chain of Z parameters from \code{BFMMM_Nu_Z_multiple_try}}

\item{nu_samp}{Cube containing initial chain of nu parameters from \code{BFMMM_Nu_Z_multiple_try}}

\item{eta_samp}{Field of Cubes containing inital chain of eta parameters from \code{BFMMM_Nu_Z_multiple_try}}

\item{burnin_prop}{Double containing proportion of chain used to estimate the starting point of nu parameters and Z parameters}

\item{c}{Vector containing hyperparmeter for sampling from pi (If left NULL, the one vector will be used)}

\item{b}{double containing hyperparamete for sampling from alpha_3}

\item{nu_1}{double containing hyperparameter for sampling from gamma}

\item{alpha1l}{Double containing hyperparameter for sampling from A}

\item{alpha2l}{Double containing hyperparameter for sampling from A}

\item{beta1l}{Double containing hyperparameter for sampling from A (scale)}

\item{beta2l}{Double containing hyperparameter for sampling from A (scale)}

\item{a_Z_PM}{Double containing hyperparameter of the random walk MH for Z parameter}

\item{a_pi_PM}{Double containing hyperparameter of the random walk MH for pi parameter}

\item{var_alpha3}{Double containing variance parameter of the random walk MH for alpha_3 parameter}

\item{var_epsilon1}{Double containing hyperparameter for sampling from A having to do with variance for Metropolis-Hastings algorithm}

\item{var_epsilon2}{Double containing hyperparameter for sampling from A having to do with variance for Metropolis-Hastings algorithm}

\item{alpha_nu}{Double containing hyperparameter for sampling from tau_nu}

\item{beta_nu}{Double containing hyperparameter for sampling from tau_nu (scale)}

\item{alpha_eta}{Double containing hyperparameter for sampling from tau_eta}

\item{beta_eta}{Double containing hyperparameter for sampling from tau_eta (scale)}

\item{alpha_0}{Double containing hyperparameter for sampling from sigma}

\item{beta_0}{Double containing hyperparameter for sampling from sigma (scale)}

\item{covariance_adj}{Boolean containing whether or not covariates should affect the covariance}
}
\value{
a List containing:
\describe{
  \item{\code{B}}{The basis functions evaluated at the observed time points}
  \item{\code{Z}}{estimates of Z}
  \item{\code{nu}}{estimates of nu}
  \item{\code{chi}}{chi samples from MCMC chain}
  \item{\code{A}}{A samples from MCMC chain}
  \item{\code{delta}}{delta samples from MCMC chain}
  \item{\code{sigma}}{sigma samples from MCMC chain}
  \item{\code{tau}}{tau samples from MCMC chain}
  \item{\code{gamma}}{gamma samples from the MCMC chain}
  \item{\code{Phi}}{Phi samples from MCMC chain}
  \item{\code{loglik}}{Log-likelihood plot of best performing chain}
}
}
\description{
This function is meant to be used after using \code{BFMMM_NU_Z_multiple_try}.
This function samples from the rest of the model parameters given a fixed value of
nu and Z. The fixed value of nu and Z are found by using the best markov chain
found in \code{BFMMM_NU_Z_multiple_try}. Once this function is ran, the results
can be used in \code{BFMMM_warm_start}.
}
\section{Warning}{

The following must be true:
\describe{
  \item{\code{tot_mcmc_iters}}{must be an integer larger than or equal to 100}
  \item{\code{n_try}}{must be an integer larger than or equal to 1}
  \item{\code{burnin_prop}}{must be between 0 and 1}
  \item{\code{k}}{must be an integer larger than or equal to 2}
  \item{\code{n_funct}}{must be an integer larger than 1}
  \item{\code{basis_degree}}{must be an integer larger than or equal to 1}
  \item{\code{n_eigen}}{must be greater than or equal to 1}
  \item{\code{internal_knots}}{must lie in the range of \code{boundary_knots}}
  \item{\code{c}}{must be greater than 0 and have k elements}
  \item{\code{b}}{must be positive}
  \item{\code{nu_1}}{must be positive}
  \item{\code{alpha1l}}{must be positive}
  \item{\code{beta1l}}{must be positive}
  \item{\code{alpha2l}}{must be positive}
  \item{\code{beta1l}}{must be positive}
  \item{\code{a_Z_PM}}{must be positive}
  \item{\code{a_pi_PM}}{must be positive}
  \item{\code{var_alpha3}}{must be positive}
  \item{\code{var_epsilon1}}{must be positive}
  \item{\code{var_epsilon2}}{must be positive}
  \item{\code{alpha}}{must be positive}
  \item{\code{beta}}{must be positive}
  \item{\code{alpha_0}}{must be positive}
  \item{\code{beta_0}}{must be positive}
}
}

\examples{
## Load sample data
Y <- readRDS(system.file("test-data", "Sim_data.RDS", package = "BayesFMMM"))
time <- readRDS(system.file("test-data", "time.RDS", package = "BayesFMMM"))

## Set Hyperparameters
tot_mcmc_iters <- 150
n_try <- 1
k <- 2
n_funct <- 40
basis_degree <- 3
n_eigen <- 3
boundary_knots <- c(0, 1000)
internal_knots <- c(250, 500, 750)

X <- matrix(rnorm(40, 0 , 1), nrow = 40, ncol = 1)
## Get Estimates of Z and nu
est1 <- BFMMM_Nu_Z_multiple_try_Cov_Adj(tot_mcmc_iters, n_try, k, Y, time, X, n_funct,
                                        basis_degree, n_eigen, boundary_knots,
                                        internal_knots)

## Run function
est2 <- BFMMM_Theta_est_Cov_Adj(tot_mcmc_iters, n_try, k, Y, time, X, n_funct,
                                basis_degree, n_eigen, boundary_knots,
                                internal_knots, est1$Z, est1$nu, est1$eta)

}
