% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BFOC}
\alias{BFOC}
\title{Run the Bayesian Functional Overlapping Clusters model}
\usage{
BFOC(
  y_obs,
  t_obs,
  n_funct,
  K,
  P,
  M,
  tot_mcmc_iters,
  t_star,
  nu_1,
  rho,
  alpha_3,
  a_12,
  alpha1l,
  alpha2l,
  beta1l,
  beta2l,
  var_epsilon1,
  var_epsilon2,
  alpha,
  beta,
  alpha_0,
  beta_0
)
}
\arguments{
\item{y_obs}{Field (list) of vectors containing the observed values}

\item{t_obs}{Field (list) of vectors containing time points of observed values}

\item{n_funct}{Double containing number of functions observed}

\item{P}{Int that indicates the number of b-spline basis functions}

\item{M}{int that indicates the number of slices used in Phi parameter}

\item{tot_mcmc_iters}{Int containing total number of MCMC iterations}

\item{t_star}{Field (list) of vectors containing time points of interest that are not observed (optional)}

\item{rho}{Double containing hyperparmater for sampling from Z}

\item{alpha_3}{Double hyperparameter for sampling from pi}

\item{a_12}{Vec containing hyperparameters for sampling from delta}

\item{alpha1l}{Double containing hyperparameters for sampling from A}

\item{alpha2l}{Double containing hyperparameters for sampling from A}

\item{beta1l}{Double containing hyperparameters for sampling from A}

\item{beta2l}{Double containing hyperparameters for sampling from A}

\item{alpha}{Double containing hyperparameters for sampling from tau}

\item{beta}{Double containing hyperparameters for sampling from tau}

\item{alpha_0}{Double containing hyperparameters for sampling from sigma}

\item{beta_0}{Double containing hyperparameters for sampling from sigma}

\item{var_epslion1}{Double containing hyperparameters for sampling from A having to do with variance for Metropolis-Hastings algorithm}

\item{var_epslion2}{Double containing hyperparameters for sampling from A having to do with variance for Metropolis-Hastings algorithm}
}
\description{
Run the Bayesian Functional Overlapping Clusters model
}
