% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BFOC_U_TT}
\alias{BFOC_U_TT}
\alias{BFOC_U_Temp}
\title{Conducts tempered MCMC to estimate the posterior distribution in an unsupervised setting. MCMC samples will be stored in batches to a specified path.}
\usage{
BFOC_U_TT(
  y_obs,
  t_obs,
  n_funct,
  K,
  P,
  M,
  tot_mcmc_iters,
  r_stored_iters,
  t_star,
  nu_1,
  rho,
  alpha_3,
  alpha1l,
  alpha2l,
  beta1l,
  beta2l,
  var_epsilon1,
  var_epsilon2,
  alpha,
  beta,
  alpha_0,
  beta_0,
  directory,
  Z_est,
  A_est,
  pi_est,
  tau_est,
  delta_est,
  nu_est,
  Phi_est,
  gamma_est,
  chi_est,
  y_star_est,
  beta_N_t,
  N_t,
  sigma_est
)

BFOC_U_Temp(
  y_obs,
  t_obs,
  n_funct,
  K,
  P,
  M,
  tot_mcmc_iters,
  r_stored_iters,
  t_star,
  nu_1,
  rho,
  alpha_3,
  alpha1l,
  alpha2l,
  beta1l,
  beta2l,
  var_epsilon1,
  var_epsilon2,
  alpha,
  beta,
  alpha_0,
  beta_0,
  directory,
  Z_est,
  A_est,
  pi_est,
  tau_est,
  delta_est,
  nu_est,
  Phi_est,
  gamma_est,
  chi_est,
  y_star_est,
  temp,
  sigma_est
)
}
\arguments{
\item{y_obs}{Field (list) of vectors containing the observed values}

\item{t_obs}{Field (list) of vectors containing time points of observed values}

\item{n_funct}{Double containing number of functions observed}

\item{P}{Int that indicates the number of b-spline basis functions}

\item{M}{int that indicates the number of slices used in Phi parameter}

\item{tot_mcmc_iters}{Int containing total number of MCMC iterations}

\item{r_stored_iters}{Int constaining number of iterations performed for each batch}

\item{t_star}{Field (list) of vectors containing time points of interest that are not observed (optional)}

\item{rho}{Double containing hyperparmater for sampling from Z}

\item{alpha_3}{Double hyperparameter for sampling from pi}

\item{alpha1l}{Double containing hyperparameters for sampling from A}

\item{alpha2l}{Double containing hyperparameters for sampling from A}

\item{beta1l}{Double containing hyperparameters for sampling from A}

\item{beta2l}{Double containing hyperparameters for sampling from A}

\item{alpha}{Double containing hyperparameters for sampling from tau}

\item{beta}{Double containing hyperparameters for sampling from tau}

\item{alpha_0}{Double containing hyperparameters for sampling from sigma}

\item{beta_0}{Double containing hyperparameters for sampling from sigma}

\item{directory}{String containing path to store batches of MCMC samples}

\item{Z_est}{Matrix containing initial starting point of Z matrix}

\item{A_est}{Vector containing initial starting point of A vector}

\item{pi_est}{Vector containing initial starting point of pi vector}

\item{tau_est}{Vector containing initial starting point of tau vector}

\item{delta_est}{Vector containing initial starting point of delta vector}

\item{nu_est}{Matrix containing initial starting point of nu matrix}

\item{Phi_est}{Cube containing initial starting point of Phi matrix}

\item{gamma_est}{Cube containing initial starting point of gamma matrix}

\item{chi_est}{Matrix containing initial starting point of chi matrix}

\item{y_star_est}{Field of Vectors containing initial starting point of y_star}

\item{sigma_est}{Double containing starting point of sigma parameter}

\item{a_12}{Vec containing hyperparameters for sampling from delta}

\item{var_epslion1}{Double containing hyperparameters for sampling from A having to do with variance for Metropolis-Hastings algorithm}

\item{var_epslion2}{Double containing hyperparameters for sampling from A having to do with variance for Metropolis-Hastings algorithm}
}
\value{
params List of objects containing the MCMC samples from the last batch

params List of objects containing the MCMC samples from the last batch
}
\description{
Conducts tempered MCMC to estimate the posterior distribution in an unsupervised setting. MCMC samples will be stored in batches to a specified path.

Conducts tempered MCMC to estimate the posterior distribution in an unsupervised setting. MCMC samples will be stored in batches to a specified path.
}
