% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BFPMM_MTT}
\alias{BFPMM_MTT}
\title{Conducts a mixture of untempered sampling and tempered sampling to get posterior draws from the partial membership model}
\usage{
BFPMM_MTT(
  y_obs,
  t_obs,
  n_funct,
  thinning_num,
  K,
  basis_degree,
  M,
  boundary_knots,
  internal_knots,
  tot_mcmc_iters,
  r_stored_iters,
  n_temp_trans,
  c,
  b,
  nu_1,
  alpha1l,
  alpha2l,
  beta1l,
  beta2l,
  a_Z_PM,
  a_pi_PM,
  var_alpha3,
  var_epsilon1,
  var_epsilon2,
  alpha,
  beta,
  alpha_0,
  beta_0,
  directory,
  beta_N_t,
  N_t
)
}
\arguments{
\item{y_obs}{Field (list) of vectors containing the observed values}

\item{t_obs}{Field (list) of vectors containing time points of observed values}

\item{n_funct}{Int containing number of functions observed}

\item{thinning_num}{Int containing how often we save an MCMC iteration}

\item{K}{Int containing the number of clusters}

\item{M}{Int containing the number of eigenfunctions}

\item{boundary_knots}{Vector containing the boundary points of our index domain of interest}

\item{internal_knots}{Vector location of internal knots for B-splines}

\item{tot_mcmc_iters}{Int containing total number of MCMC iterations}

\item{r_stored_iters}{Int constaining number of iterations performed for each batch}

\item{alpha1l}{Double containing hyperparameters for sampling from A}

\item{alpha2l}{Double containing hyperparameters for sampling from A}

\item{beta1l}{Double containing hyperparameters for sampling from A}

\item{beta2l}{Double containing hyperparameters for sampling from A}

\item{a_Z_PM}{Double containing hyperparameter used to sample from the posterior of Z}

\item{a_pi_PM}{Double containing hyperparameter used to sample from the posterior of pi}

\item{var_alpha3}{Double containing hyperparameter for sampling from alpha_3}

\item{alpha}{Double containing hyperparameters for sampling from tau}

\item{beta}{Double containing hyperparameters for sampling from tau}

\item{alpha_0}{Double containing hyperparameters for sampling from sigma}

\item{beta_0}{Double containing hyperparameters for sampling from sigma}

\item{directory}{String containing path to store batches of MCMC samples}

\item{basis}{degree Int containing the degree of B-splines used}

\item{rho}{Double containing hyperparmater for sampling from Z}

\item{alpha_3}{Double hyperparameter for sampling from pi}

\item{a_12}{Vec containing hyperparameters for sampling from delta}

\item{var_epslion1}{Double containing hyperparameters for sampling from A having to do with variance for Metropolis-Hastings algorithm}

\item{var_epslion2}{Double containing hyperparameters for sampling from A having to do with variance for Metropolis-Hastings algorithm}
}
\value{
params List of objects containing the MCMC samples from the last batch
}
\description{
Conducts a mixture of untempered sampling and tempered sampling to get posterior draws from the partial membership model
}
