% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BFPMM_Nu_Z_multiple_try}
\alias{BFPMM_Nu_Z_multiple_try}
\title{Function for finding a good initial starting point for nu parameters and Z parameters, with option for temperered transitions}
\usage{
BFPMM_Nu_Z_multiple_try(
  tot_mcmc_iters,
  n_try,
  k,
  Y,
  time,
  n_funct,
  n_basis,
  n_eigen
)
}
\arguments{
\item{tot_mcmc_iters}{Int conatining the number of MCMC iterations per try}

\item{n_try}{Int containing how many different chains are tried}

\item{k}{Int containing the number of clusters}

\item{Y}{Field of vectors containing the observed values}

\item{time}{Field of vecotrs containing the observed time points}

\item{n_funct}{Int containing the number of functions}

\item{n_basis}{Int containing the number of basis functions}

\item{n_eigen}{Int containing the number of eigenfunctions}
}
\value{
BestChain List containing a summary of the best performing chain
}
\description{
Function for finding a good initial starting point for nu parameters and Z parameters, with option for temperered transitions
}
