% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BFPMM_Nu_ZMV}
\alias{BFPMM_Nu_ZMV}
\title{Conducts un-tempered MCMC to mean and allocation parameters for the multivariate model}
\usage{
BFPMM_Nu_ZMV(
  y_obs,
  K,
  M,
  tot_mcmc_iters,
  c,
  b,
  nu_1,
  alpha1l,
  alpha2l,
  beta1l,
  beta2l,
  a_Z_PM,
  a_pi_PM,
  var_alpha3,
  var_epsilon1,
  var_epsilon2,
  alpha,
  beta,
  alpha_0,
  beta_0
)
}
\arguments{
\item{y_obs}{Matrix of observed vectors}

\item{K}{Int containing the number of clusters}

\item{M}{Int containing the number of eigenvectors}

\item{tot_mcmc_iters}{Int containing total number of MCMC iterations}

\item{c}{Vector containing hyperparmeters for pi}

\item{b}{double containing hyperparameter for alpha_3}

\item{alpha1l}{Double containing hyperparameters for sampling from A}

\item{alpha2l}{Double containing hyperparameters for sampling from A}

\item{beta1l}{Double containing hyperparameters for sampling from A}

\item{beta2l}{Double containing hyperparameters for sampling from A}

\item{var_alpha3}{Double containing variance parameter of the random walk MH for alpha_3 parameter}

\item{alpha}{Double containing hyperparameters for sampling from tau}

\item{beta}{Double containing hyperparameters for sampling from tau}

\item{alpha_0}{Double containing hyperparameters for sampling from sigma}

\item{beta_0}{Double containing hyperparameters for sampling from sigma}

\item{r_stored_iters}{Int containing number of iterations performed for each batch}

\item{a_12}{Vector containing hyperparameters for sampling from delta}

\item{var_pi}{Double containing variance parameter of the random walk MH for pi parameter}

\item{var_Z}{Double containing variance parameter of the random walk MH for Z parameter}

\item{var_epslion1}{Double containing hyperparameters for sampling from A having to do with variance for Metropolis-Hastings algorithm}

\item{var_epslion2}{Double containing hyperparameters for sampling from A having to do with variance for Metropolis-Hastings algorithm}

\item{directory}{String containing path to store batches of MCMC samples}
}
\value{
params List of objects containing the MCMC samples from the last batch
}
\description{
Conducts un-tempered MCMC to mean and allocation parameters for the multivariate model
}
