% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BFPMM_Templadder}
\alias{BFPMM_Templadder}
\title{Conducts tempered MCMC to estimate the posterior distribution in an unsupervised setting. MCMC samples will be stored in batches to a specified path.}
\usage{
BFPMM_Templadder(
  y_obs,
  t_obs,
  n_funct,
  K,
  P,
  M,
  tot_mcmc_iters,
  r_stored_iters,
  c,
  b,
  nu_1,
  alpha1l,
  alpha2l,
  beta1l,
  beta2l,
  a_Z_PM,
  a_pi_PM,
  var_alpha3,
  var_epsilon1,
  var_epsilon2,
  alpha,
  beta,
  alpha_0,
  beta_0,
  beta_N_t,
  N_t
)
}
\arguments{
\item{y_obs}{Field (list) of vectors containing the observed values}

\item{t_obs}{Field (list) of vectors containing time points of observed values}

\item{n_funct}{Double containing number of functions observed}

\item{P}{Int that indicates the number of b-spline basis functions}

\item{M}{int that indicates the number of slices used in Phi parameter}

\item{tot_mcmc_iters}{Int containing total number of MCMC iterations}

\item{r_stored_iters}{Int constaining number of iterations performed for each batch}

\item{alpha1l}{Double containing hyperparameters for sampling from A}

\item{alpha2l}{Double containing hyperparameters for sampling from A}

\item{beta1l}{Double containing hyperparameters for sampling from A}

\item{beta2l}{Double containing hyperparameters for sampling from A}

\item{a_Z_PM}{Double containing hyperparameter used to sample from the posterior of Z}

\item{a_pi_PM}{Double containing hyperparameter used to sample from the posterior of pi}

\item{var_alpha3}{Doubel containing hyperparameter for sampling from alpha_3}

\item{alpha}{Double containing hyperparameters for sampling from tau}

\item{beta}{Double containing hyperparameters for sampling from tau}

\item{alpha_0}{Double containing hyperparameters for sampling from sigma}

\item{beta_0}{Double containing hyperparameters for sampling from sigma}

\item{rho}{Double containing hyperparmater for sampling from Z}

\item{alpha_3}{Double hyperparameter for sampling from pi}

\item{a_12}{Vec containing hyperparameters for sampling from delta}

\item{var_epslion1}{Double containing hyperparameters for sampling from A having to do with variance for Metropolis-Hastings algorithm}

\item{var_epslion2}{Double containing hyperparameters for sampling from A having to do with variance for Metropolis-Hastings algorithm}

\item{directory}{String containing path to store batches of MCMC samples}
}
\value{
params List of objects containing the MCMC samples from the last batch
}
\description{
Conducts tempered MCMC to estimate the posterior distribution in an unsupervised setting. MCMC samples will be stored in batches to a specified path.
}
