% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BMFPMM_Nu_Z_multiple_try}
\alias{BMFPMM_Nu_Z_multiple_try}
\title{Function for finding a good initial starting point for nu parameters and Z parameters for multivariate functional data, with option for temperered transitions}
\usage{
BMFPMM_Nu_Z_multiple_try(
  tot_mcmc_iters,
  n_try,
  k,
  Y,
  time,
  n_funct,
  basis_degree,
  n_eigen,
  boundary_knots,
  internal_knots
)
}
\arguments{
\item{tot_mcmc_iters}{Int containing the number of MCMC iterations per try}

\item{n_try}{Int containing how many different chains are tried}

\item{k}{Int containing the number of clusters}

\item{Y}{Field of vectors containing the observed values}

\item{time}{field of matrices that contain the observed time points (each column is a dimension)}

\item{n_funct}{Int containing the number of functions}

\item{basis_degree}{vector containing the desired basis degree for each dimension}

\item{n_eigen}{Int containing the number of eigenfunctions}

\item{boundary_knots}{matrix containing the boundary knots for each dimension (each row is a dimension)}

\item{internal_knots}{field of vectors containing the internal knots for each dimension}
}
\value{
BestChain List containing a summary of the best performing chain
}
\description{
Function for finding a good initial starting point for nu parameters and Z parameters for multivariate functional data, with option for temperered transitions
}
