% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BMFPMM_Theta_Est}
\alias{BMFPMM_Theta_Est}
\title{Estimates the initial starting point of the rest of the parameters given an initial starting point for Z and nu for multivariate functional data}
\usage{
BMFPMM_Theta_Est(
  tot_mcmc_iters,
  Z_samp,
  nu_samp,
  burnin_prop,
  k,
  Y,
  time,
  n_funct,
  basis_degree,
  n_eigen,
  boundary_knots,
  internal_knots
)
}
\arguments{
\item{tot_mcmc_iters}{Int containing the total number of MCMC iterations}

\item{Z_samp}{Cube containing initial chain of Z parameters from BFPMM_Nu_Z_multiple_try}

\item{nu_samp}{Cube containing initial chain of nu parameters from BFPMM_Nu_Z_multiple_try}

\item{burnin_prop}{Double containing proportion of chain used to estimate the starting point of nu parameters and Z parameters}

\item{k}{Int containing the number of clusters}

\item{Y}{Field of vectors containing the observed values}

\item{time}{Field of vectors containing the observed time points}

\item{n_funct}{Int containing the number of functions}

\item{basis_degree}{Int containing the degree of B-splines used}

\item{n_eigen}{Int containing the number of eigenfunctions}

\item{boundary_knots}{Vector containing the boundary points of our index domain of interest}

\item{internal_knots}{Vector location of internal knots for B-splines}
}
\value{
BestChain List containing a summary of the chain conditioned on nu and Z
}
\description{
Estimates the initial starting point of the rest of the parameters given an initial starting point for Z and nu for multivariate functional data
}
