% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BMVMMM_Theta_est}
\alias{BMVMMM_Theta_est}
\title{Find initial starting points for covariance parameters given mean and allocation structure
for multivariate data}
\usage{
BMVMMM_Theta_est(
  tot_mcmc_iters,
  n_try,
  K,
  Y,
  n_eigen,
  multiple_try,
  X = NULL,
  burnin_prop = 0.8,
  c = NULL,
  b = 10,
  nu_1 = 3,
  alpha1l = 1,
  alpha2l = 2,
  beta1l = 1,
  beta2l = 1,
  a_Z_PM = 10000,
  a_pi_PM = 1000,
  var_alpha3 = 0.05,
  var_epsilon1 = 1,
  var_epsilon2 = 1,
  alpha_nu = 10,
  beta_nu = 1,
  alpha_eta = 10,
  beta_eta = 1,
  alpha_0 = 1,
  beta_0 = 1,
  covariance_adj = FALSE
)
}
\arguments{
\item{tot_mcmc_iters}{Int containing the total number of MCMC iterations}

\item{n_try}{Int containing how many different chains are tried}

\item{K}{Int containing the number of clusters}

\item{Y}{Matrix of observed vectors (each row is an observation)}

\item{n_eigen}{Int containing the number of eigenfunctions}

\item{multiple_try}{List containing results from \code{BMVMMM_Nu_Z_multiple_try}}

\item{X}{Matrix of covariates, where each row corresponds to an observation (if covariate adjusted)}

\item{burnin_prop}{Double containing proportion of chain used to estimate the starting point of nu parameters and Z parameters}

\item{c}{Vector containing hyperparmeter for sampling from pi (If left NULL, the one vector will be used)}

\item{b}{double containing hyperparamete for sampling from alpha_3}

\item{nu_1}{double containing hyperparameter for sampling from gamma}

\item{alpha1l}{Double containing hyperparameter for sampling from A}

\item{alpha2l}{Double containing hyperparameter for sampling from A}

\item{beta1l}{Double containing hyperparameter for sampling from A (scale)}

\item{beta2l}{Double containing hyperparameter for sampling from A (scale)}

\item{a_Z_PM}{Double containing hyperparameter of the random walk MH for Z parameter}

\item{a_pi_PM}{Double containing hyperparameter of the random walk MH for pi parameter}

\item{var_alpha3}{Double containing variance parameter of the random walk MH for alpha_3 parameter}

\item{var_epsilon1}{Double containing hyperparameter for sampling from A having to do with variance for Metropolis-Hastings algorithm}

\item{var_epsilon2}{Double containing hyperparameter for sampling from A having to do with variance for Metropolis-Hastings algorithm}

\item{alpha_nu}{Double containing hyperparameter for sampling from tau_nu}

\item{beta_nu}{Double containing hyperparameter for sampling from tau_nu (scale)}

\item{alpha_eta}{Double containing hyperparameter for sampling from tau_eta}

\item{beta_eta}{Double containing hyperparameter for sampling from tau_eta (scale)}

\item{alpha_0}{Double containing hyperparameter for sampling from sigma}

\item{beta_0}{Double containing hyperparameter for sampling from sigma (scale)}

\item{covariance_adj}{Boolean containing whether or not covariates should affect the covariance}
}
\value{
a List containing:
\describe{
  \item{\code{Z}}{estimates of Z}
  \item{\code{nu}}{estimates of nu}
  \item{\code{eta}}{estimated value from \code{BFMMM_Nu_Z_multiple_try} (if covariate adjusted)}
  \item{\code{chi}}{chi samples from MCMC chain}
  \item{\code{A}}{A samples from MCMC chain}
  \item{\code{A_xi}}{A_xi samples from MCMC chain (if covariate adjusted)}
  \item{\code{delta}}{delta samples from MCMC chain}
  \item{\code{delta_xi}}{delta_xi samples from MCMC chain (if covariate adjusted)}
  \item{\code{gamma_xi}}{gamma_xi samples from MCMC chain (if covariate adjusted)}
  \item{\code{sigma}}{sigma samples from MCMC chain}
  \item{\code{tau}}{tau samples from MCMC chain}
  \item{\code{tau_eta}}{tau_eta samples from MCMC chain (if covariate adjusted)}
  \item{\code{gamma}}{gamma samples from the MCMC chain}
  \item{\code{Phi}}{Phi samples from MCMC chain}
  \item{\code{loglik}}{Log-likelihood plot of best performing chain}
}
}
\description{
This function is meant to be used after using \code{BMVMMM_NU_Z_multiple_try}.
This function samples from the rest of the model parameters given a fixed value of
nu and Z. The fixed value of nu and Z are found by using the best markov chain
found in \code{BMVMMM_NU_Z_multiple_try}. Once this function is ran, the results
can be used in \code{BMVMMM_warm_start}.To specify a covariate
adjusted model, please provide the design matrix X.
}
\section{Warning}{

The following must be true:
\describe{
  \item{\code{tot_mcmc_iters}}{must be an integer larger than or equal to 100}
  \item{\code{n_try}}{must be an integer larger than or equal to 1}
  \item{\code{K}}{must be an integer larger than or equal to 2}
  \item{\code{n_eigen}}{must be greater than or equal to 1}
  \item{\code{X}}{must have the same number of rows as Y (if specified)}
  \item{\code{burnin_prop}}{must be between 0 and 1}
  \item{\code{c}}{must be greater than 0 and have K elements}
  \item{\code{b}}{must be positive}
  \item{\code{nu_1}}{must be positive}
  \item{\code{alpha1l}}{must be positive}
  \item{\code{beta1l}}{must be positive}
  \item{\code{alpha2l}}{must be positive}
  \item{\code{beta1l}}{must be positive}
  \item{\code{a_Z_PM}}{must be positive}
  \item{\code{a_pi_PM}}{must be positive}
  \item{\code{var_alpha3}}{must be positive}
  \item{\code{var_epsilon1}}{must be positive}
  \item{\code{var_epsilon2}}{must be positive}
  \item{\code{alpha_nu}}{must be positive}
  \item{\code{beta_nu}}{must be positive}
  \item{\code{alpha_eta}}{must be positive}
  \item{\code{beta_eta}}{must be positive}
  \item{\code{alpha_0}}{must be positive}
  \item{\code{beta_0}}{must be positive}
}
}

\examples{
#########################
### Not Covariate Adj ###
#########################

## Load sample data
Y <- readRDS(system.file("test-data", "MVSim_data.RDS", package = "BayesFMMM"))

## Set Hyperparameters
tot_mcmc_iters <- 150
n_try <- 1
K <- 2
n_eigen <- 2

## Get Estimates of Z and nu
est1 <- BFMMM_Nu_Z_multiple_try(tot_mcmc_iters, n_try, K, Y, n_eigen)

## Run function
est2 <- BFMMM_Theta_est(tot_mcmc_iters, n_try, K, Y,
                        n_eigen, est1)

#####################
### Covariate Adj ###
#####################

## Load sample data
Y <- readRDS(system.file("test-data", "MVSim_data.RDS", package = "BayesFMMM"))

## Set Hyperparameters
tot_mcmc_iters <- 150
n_try <- 1
K <- 2
n_eigen <- 2

X <- matrix(rnorm(40, 0 , 1), nrow = 40, ncol = 1)

## Get Estimates of Z and nu
est1 <- BFMMM_Nu_Z_multiple_try(tot_mcmc_iters, n_try, K, Y, n_eigen, X = X)

## Run function
est2 <- BFMMM_Theta_est(tot_mcmc_iters, n_try, K, Y,
                        n_eigen, est1, X = X)

#####################################################################
### Covariate Adj  (with Covariate-depenent covariance structure) ###
#####################################################################

## Load sample data
Y <- readRDS(system.file("test-data", "MVSim_data.RDS", package = "BayesFMMM"))

## Set Hyperparameters
tot_mcmc_iters <- 150
n_try <- 1
K <- 2
n_eigen <- 2

X <- matrix(rnorm(40, 0 , 1), nrow = 40, ncol = 1)

## Get Estimates of Z and nu
est1 <- BFMMM_Nu_Z_multiple_try(tot_mcmc_iters, n_try, K, Y, n_eigen, X = X)

## Run function
est2 <- BFMMM_Theta_est(tot_mcmc_iters, n_try, K, Y,
                        n_eigen, est1, X = X, covariance_adj = T)

}
