% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_lB}
\alias{calc_lB}
\title{Calculates the log of B(a) function used in the dirichlet distribution}
\usage{
calc_lB(alpha)
}
\arguments{
\item{alpha}{Vector containing input to the function}
}
\value{
log_B Double containing the log of the output of the B function
}
\description{
Calculates the log of B(a) function used in the dirichlet distribution
}
