% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_m}
\alias{compute_m}
\title{Computes m for all observations}
\arguments{
\item{map}{Map that contains mapping for covariance matrix}

\item{S_obs}{Field of Matrices containing basis functions evaluated at observed time points}

\item{S_star}{Field of Matrices containing basis functions evaluated at unobserved time points}

\item{Z}{Matrix of current Z parameter}

\item{phi}{Matrix of current phi paramaters}

\item{nu}{Matrix containing the mean vectors as columns}

\item{i}{Int indicating which M we are calculating}

\item{mp_inv}{Field of Matrices acting as a placeholder for mp inverse}

\item{mean_ph_obs}{Field of Vectors acting as placeholder for mean of f_i at observed time points}

\item{mean_ph_star}{Field of Vectors acting as placeholder for vector with length of unobserved time points}

\item{m}{Field of Vectors acting as a placeholder for m}
}
\description{
Computes m for all observations

Computes m for all observations under common variance model
}
