% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_mi}
\alias{compute_mi}
\title{Computes m_i}
\usage{
compute_mi(S_obs, Z, phi, nu, i, m)
}
\arguments{
\item{S_obs}{Field of Matrices containing basis functions evaluated at observed time points}

\item{Z}{Matrix of current Z parameter}

\item{phi}{Cube of current phi paramaters}

\item{nu}{Matrix of current nu parameters}

\item{i}{Int indicating which M we are calculating}

\item{m}{vector acting as a placeholder for m}
}
\description{
Computes m_i
}
