% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_mi_Mi}
\alias{compute_mi_Mi}
\title{Computes m_i and M_i}
\arguments{
\item{Rho}{Matrix with each row containing the elements of the upper triangular matrix}

\item{nu}{Matrix that contains mean vectors as columns}

\item{Cov}{Matrix containing placeholder for covariance matrix}

\item{S_obs}{Field of Matrices containing basis functions evaluated at observed time points}

\item{S_star}{Field of Matrices containing basis functions evaluated at unobserved time points}

\item{Z}{Matrix of current Z parameter}

\item{Phi}{Matrix of current Phi paramaters}

\item{i}{Int indicating which M we are calculating}

\item{mp_inv}{Matrix acting as a placeholder for mp inverse}

\item{mean_ph_obs}{vector acting as placeholder for mean of f_i at observed time points}

\item{mean_ph_star}{vector acting as placeholder for vector with length of unobserved time points}

\item{m}{Vector acting as a placeholder for m}

\item{M}{Matrix acting as a placeholder for M}
}
\description{
Computes m_i and M_i

Computes m_i under common covariance matrix
}
