% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeM}
\alias{computeM}
\title{Computes all M matrices}
\usage{
computeM(S_obs, Z, phi, M)
}
\arguments{
\item{S_obs}{Field of Matrices containing basis functions evaluated at observed time points}

\item{Z}{Matrix of current Z parameter}

\item{phi}{Cube of current phi paramaters}

\item{M}{Field of Matrices containing all M matrices}
}
\description{
Computes all M matrices
}
