% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeMi}
\alias{computeMi}
\title{Computes M_i}
\usage{
computeMi(S_obs, Z, phi, i, M)
}
\arguments{
\item{S_obs}{Field of Matrices containing basis functions evaluated at observed time points}

\item{Z}{Matrix of current Z parameter}

\item{phi}{Cube of current phi paramaters}

\item{i}{Int indicating which M we are calculating}

\item{M}{Matrix acting as a placeholder for M}
}
\description{
Computes M_i
}
