% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Conditional_Predictive_Ordinates}
\alias{Conditional_Predictive_Ordinates}
\title{Conditional Predictive Ordinates}
\usage{
Conditional_Predictive_Ordinates(
  dir,
  n_files,
  n_MCMC,
  basis_degree,
  boundary_knots,
  internal_knots,
  time,
  Y,
  burnin_prop = 0.2,
  X = NULL,
  cov_adj = FALSE,
  log_CPO = TRUE
)
}
\arguments{
\item{dir}{String containing the directory where the MCMC files are located}

\item{n_files}{Int containing the number of files per parameter}

\item{n_MCMC}{Int containing the number of saved MCMC iterations per file}

\item{basis_degree}{Int containing the degree of B-splines used}

\item{boundary_knots}{Vector containing the boundary points of our index domain of interest}

\item{internal_knots}{Vector location of internal knots for B-splines}

\item{time}{Field of vectors containing time points at which the function was observed}

\item{Y}{Field of vectors containing observed values of the function}

\item{burnin_prop}{Double containing proportion of MCMC samples to discard}

\item{X}{Matrix of covariates (each row contains the covariates for a single observation) (optional arugment)}

\item{cov_adj}{Boolean containing whether the model fit had a covariance structure that is covariate-dependent (optional argument)}

\item{log_CPO}{Boolean conatining whether or not CPO is returned on the log scale (optional argument)}
}
\value{
CPO Vector containing the CPO for each observation
}
\description{
Calculates the Conditional Predictive Ordinates for functional models.
This function can handle covariate adjusted models as well as non-adjusted models.
}
