% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ConditionalPredictiveOrdinates}
\alias{ConditionalPredictiveOrdinates}
\title{Conditional Predictive Ordinates}
\usage{
ConditionalPredictiveOrdinates(
  dir,
  n_files,
  basis_degree,
  boundary_knots,
  internal_knots,
  time,
  Y,
  burnin_prop = 0.1,
  X = NULL,
  cov_adj = FALSE,
  log_CPO = TRUE
)
}
\arguments{
\item{dir}{String containing the directory where the MCMC files are located}

\item{n_files}{Int containing the number of files per parameter}

\item{basis_degree}{Int containing the degree of B-splines used}

\item{boundary_knots}{Vector containing the boundary points of our index domain of interest}

\item{internal_knots}{Vector location of internal knots for B-splines}

\item{time}{Field of vectors containing time points at which the function was observed}

\item{Y}{Field of vectors containing observed values of the function}

\item{burnin_prop}{Double containing proportion of MCMC samples to discard}

\item{X}{Matrix of covariates (each row contains the covariates for a single observation) (optional arugment)}

\item{cov_adj}{Boolean containing whether the model fit had a covariance structure that is covariate-dependent (optional argument)}

\item{log_CPO}{Boolean conatining whether or not CPO is returned on the log scale (optional argument)}
}
\value{
CPO Vector containing the CPO for each observation

//' @section Warning:
The following must be true:
\describe{
  \item{\code{n_files}}{must be an integer larger than or equal to 1}
  \item{\code{basis_degree}}{must be an integer larger than or equal to 1}
  \item{\code{internal_knots}}{must lie in the range of \code{boundary_knots}}
  \item{\code{burnin_prop}}{must be less than 1 and greater than or equal to 0}
  \item{\code{X}}{must have the same number of columns as covariates in the model (D)}
}
}
\description{
Calculates the Conditional Predictive Ordinates for functional models.
This function can handle covariate adjusted models as well as non-adjusted models.
}
\examples{
#########################
### Not Covariate Adj ###
#########################

## Observed Data and corresponding time points
Y <- readRDS(system.file("test-data", "Sim_data.RDS", package = "BayesFMMM"))
time <- readRDS(system.file("test-data", "time.RDS", package = "BayesFMMM"))

## Directory of saved MCMC iterations
dir <- paste0(system.file("test-data", "Functional_trace", package = "BayesFMMM"),"/")

## Set Hyperparameters
n_files <- 1
basis_degree <- 3
boundary_knots <- c(0, 1000)
internal_knots <- c(250, 500, 750)

## Get CPO
CPO <- ConditionalPredictiveOrdinates(dir, n_files, basis_degree, boundary_knots,
                                      internal_knots, time, Y)

#####################
### Covariate Adj ###
#####################

## Observed Data, corresponding time points, and observed covariates
Y <- readRDS(system.file("test-data", "Sim_data.RDS", package = "BayesFMMM"))
time <- readRDS(system.file("test-data", "time.RDS", package = "BayesFMMM"))
X <- matrix(rnorm(40), ncol = 1)

## Directory of saved MCMC iterations
dir <- paste0(system.file("test-data", "Functional_trace", package = "BayesFMMM"),"/")

## Set Hyperparameters
n_files <- 1
basis_degree <- 3
boundary_knots <- c(0, 1000)
internal_knots <- c(250, 500, 750)

## Get CPO
CPO <- ConditionalPredictiveOrdinates(dir, n_files, basis_degree, boundary_knots,
                                      internal_knots, time, Y, X = X)

#####################################################################
### Covariate Adj  (with Covariate-depenent covariance structure) ###
#####################################################################

## Observed Data, corresponding time points, and observed covariates
Y <- readRDS(system.file("test-data", "Sim_data.RDS", package = "BayesFMMM"))
time <- readRDS(system.file("test-data", "time.RDS", package = "BayesFMMM"))
X <- matrix(rnorm(40), ncol = 1)

## Directory of saved MCMC iterations
dir <- paste0(system.file("test-data", "Functional_trace", package = "BayesFMMM"),"/")

## Set Hyperparameters
n_files <- 1
basis_degree <- 3
boundary_knots <- c(0, 1000)
internal_knots <- c(250, 500, 750)

## Get CPO
CPO <- ConditionalPredictiveOrdinates(dir, n_files, basis_degree, boundary_knots,
                                      internal_knots, time, Y, X = X,
                                      cov_adj = TRUE)

}
