% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convert_Z_tilde_Z}
\alias{convert_Z_tilde_Z}
\title{Converts from the transformed space to the original parameter space}
\usage{
convert_Z_tilde_Z(Z_tilde, Z)
}
\arguments{
\item{Z_tilde}{Row Vector containing parameters in the transformed space}

\item{Z}{Row Vector containing placeholder for variables in the untransformed space}
}
\description{
Converts from the transformed space to the original parameter space
}
