% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FAIC}
\alias{FAIC}
\title{Calculates the AIC of a functional model}
\usage{
FAIC(
  dir,
  n_files,
  basis_degree,
  boundary_knots,
  internal_knots,
  time,
  Y,
  burnin_prop = 0.2,
  X = NULL,
  cov_adj = FALSE
)
}
\arguments{
\item{dir}{String containing the directory where the MCMC files are located}

\item{n_files}{Int containing the number of files per parameter}

\item{basis_degree}{Int containing the degree of B-splines used}

\item{boundary_knots}{Vector containing the boundary points of our index domain of interest}

\item{internal_knots}{Vector location of internal knots for B-splines}

\item{time}{Field of vectors containing time points at which the function was observed}

\item{Y}{Field of vectors containing observed values of the function}

\item{burnin_prop}{Double containing proportion of MCMC samples to discard}

\item{X}{Matrix of covariates, where each row corresponds to an observation (if covariate adjusted)}

\item{cov_adj}{Boolean containing whether or not the covariance structure depends on the covariates of interest}
}
\value{
AIC Double containing AIC value
}
\description{
Calculates the AIC of a functional model
}
