% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FCovCI}
\alias{FCovCI}
\title{Calculates the credible interval for the covariance (Functional Data)}
\usage{
FCovCI(
  dir,
  n_files,
  n_MCMC,
  time1,
  time2,
  basis_degree,
  boundary_knots,
  internal_knots,
  l,
  m,
  alpha = 0.05,
  rescale = TRUE,
  simultaneous = FALSE
)
}
\arguments{
\item{dir}{String containing the directory where the MCMC files are located}

\item{n_files}{Int containing the number of files per parameter}

\item{n_MCMC}{Int containing the number of saved MCMC iterations per file}

\item{time1}{Vector containing time points of interest for first cluster}

\item{time2}{Vector containing time points of interest for second cluster}

\item{basis_degree}{Int containing the degree of B-splines used}

\item{boundary_knots}{Vector containing the boundary points of our index domain of interest}

\item{internal_knots}{Vector location of internal knots for B-splines}

\item{l}{Int containing the 1st cluster group of which you want to get the credible interval for}

\item{m}{Int containing the 2nd cluster group of which you want to get the credible interval for}

\item{alpha}{Double specifying the percentile of the credible interval ((1 - alpha) * 100 percent)}

\item{rescale}{Boolean indicating whether or not we should rescale the Z variables so that there is at least one observation almost completely in one group}

\item{simultaneous}{Boolean indicating whether or not the credible intervals should be simulatneous credible intevals or pointwise credible intervals}
}
\value{
CI list containing the 97.5th , 50th, and 2.5th pointwise credible functions
}
\description{
Calculates the credible interval for the covariance (Functional Data)
}
