% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FDIC}
\alias{FDIC}
\title{Calculates the DIC of a functional model}
\usage{
FDIC(
  dir,
  n_files,
  basis_degree,
  boundary_knots,
  internal_knots,
  time,
  Y,
  burnin_prop = 0.1,
  X = NULL,
  cov_adj = FALSE
)
}
\arguments{
\item{dir}{String containing the directory where the MCMC files are located}

\item{n_files}{Int containing the number of files per parameter}

\item{basis_degree}{Int containing the degree of B-splines used}

\item{boundary_knots}{Vector containing the boundary points of our index domain of interest}

\item{internal_knots}{Vector location of internal knots for B-splines}

\item{time}{Field of vectors containing time points at which the function was observed}

\item{Y}{Field of vectors containing observed values of the function}

\item{burnin_prop}{Double containing proportion of MCMC samples to discard}

\item{X}{Matrix of covariates, where each row corresponds to an observation (if covariate adjusted)}

\item{cov_adj}{Boolean containing whether or not the covariance structure depends on the covariates of interest}
}
\value{
DIC Double containing DIC value
}
\description{
Calculates the DIC of a functional model
}
\examples{
#########################
### Not Covariate Adj ###
#########################

## Observed Data and corresponding time points
Y <- readRDS(system.file("test-data", "Sim_data.RDS", package = "BayesFMMM"))
time <- readRDS(system.file("test-data", "time.RDS", package = "BayesFMMM"))

## Directory of saved MCMC iterations
dir <- paste0(system.file("test-data", "Functional_trace", package = "BayesFMMM"),"/")

## Hyperparameters
n_files <- 1
K <- 2
basis_degree <- 3
boundary_knots <- c(0, 1000)
internal_knots <- c(250, 500, 750)


## Get DIC
DIC <- FDIC(dir, n_files, basis_degree, boundary_knots, internal_knots, time, Y)

#####################
### Covariate Adj ###
#####################

## Observed data, corresponding time points, and observed covariates
Y <- readRDS(system.file("test-data", "Sim_data.RDS", package = "BayesFMMM"))
time <- readRDS(system.file("test-data", "time.RDS", package = "BayesFMMM"))
X <- matrix(rnorm(40, 0 , 1), nrow = 40, ncol = 1)

## Directory of saved MCMC iterations
dir <- paste0(system.file("test-data", "Functional_trace", package = "BayesFMMM"),"/")

## Hyperparameters
K <- 2
basis_degree <- 3
boundary_knots <- c(0, 1000)
internal_knots <- c(250, 500, 750)

## Get DIC
DIC <- FDIC(dir, n_files, basis_degree, boundary_knots, internal_knots, time,
            Y, X = X)

#####################################################################
### Covariate Adj  (with Covariate-depenent covariance structure) ###
#####################################################################

## Observed data, corresponding time points, and observed covariates
Y <- readRDS(system.file("test-data", "Sim_data.RDS", package = "BayesFMMM"))
time <- readRDS(system.file("test-data", "time.RDS", package = "BayesFMMM"))
X <- matrix(rnorm(40, 0 , 1), nrow = 40, ncol = 1)

## Directory of saved MCMC iterations
dir <- paste0(system.file("test-data", "Functional_trace", package = "BayesFMMM"),"/")

## Hyperparameters
n_files <- 1
K <- 2
basis_degree <- 3
boundary_knots <- c(0, 1000)
internal_knots <- c(250, 500, 750)

## Get DIC
DIC <- FDIC(dir, n_files, basis_degree, boundary_knots, internal_knots, time,
            Y, X = X, cov_adj = TRUE)

}
