% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FLLik}
\alias{FLLik}
\title{Calculates the Log-Likelihood of a Functional Model}
\usage{
FLLik(
  dir,
  n_files,
  basis_degree,
  boundary_knots,
  internal_knots,
  time,
  Y,
  X = NULL,
  cov_adj = FALSE
)
}
\arguments{
\item{dir}{String containing the directory where the MCMC files are located}

\item{n_files}{Int containing the number of files per parameter}

\item{basis_degree}{Int containing the degree of B-splines used}

\item{boundary_knots}{Vector containing the boundary points of our index domain of interest}

\item{internal_knots}{Vector location of internal knots for B-splines}

\item{time}{Field of vectors containing time points at which the function was observed}

\item{Y}{Field of vectors containing observed values of the function}

\item{X}{Matrix of covariates (each row contains the covariates for a single observation) (optional arugment)}

\item{cov_adj}{Boolean containing whether the model fit had a covariance structure that is covariate-dependent (optional argument)}
}
\value{
LLik Vector containing the log-likelihood evaluated at each iteration
}
\description{
Calculates the log-likelihood of the parameters for each iteration for functional models.
This function can handle covariate adjusted models as well as non-adjusted models.
}
\section{Warning}{

The following must be true:
\describe{
  \item{\code{n_files}}{must be an integer larger than or equal to 1}
  \item{\code{basis_degree}}{must be an integer larger than or equal to 1}
  \item{\code{internal_knots}}{must lie in the range of \code{boundary_knots}}
  \item{\code{X}}{must have the same number of columns as covariates in the model (D)}
}
}

\examples{
#########################
### Not Covariate Adj ###
#########################

## Set Hyperparameters
Y <- readRDS(system.file("test-data", "MVSim_data.RDS", package = "BayesFMMM"))
time <- readRDS(system.file("test-data", "time.RDS", package = "BayesFMMM"))
dir <- system.file("test-data", "Multivariate_trace", "", package = "BayesFMMM")
n_files <- 1
basis_degree <- 3
boundary_knots <- c(0, 1000)
internal_knots <- c(250, 500, 750)

## Get CI for mean function
LL <- FLLik(dir, n_files, basis_degree, boundary_knots, internal_knots,
            time, Y)

#####################
### Covariate Adj ###
#####################

## Set Hyperparameters
Y <- readRDS(system.file("test-data", "MVSim_data.RDS", package = "BayesFMMM"))
time <- readRDS(system.file("test-data", "time.RDS", package = "BayesFMMM"))
dir <- system.file("test-data", "Multivariate_trace", "", package = "BayesFMMM")
n_files <- 1
basis_degree <- 3
boundary_knots <- c(0, 1000)
internal_knots <- c(250, 500, 750)
X <- matrix(seq(-2, 2, 0.2), ncol = 1)

## Get CI for mean function
LL <- FLLik(dir, n_files, basis_degree, boundary_knots, internal_knots,
            time, Y, X = X)

#####################################################################
### Covariate Adj  (with Covariate-depenent covariance structure) ###
#####################################################################

## Set Hyperparameters
Y <- readRDS(system.file("test-data", "MVSim_data.RDS", package = "BayesFMMM"))
time <- readRDS(system.file("test-data", "time.RDS", package = "BayesFMMM"))
dir <- system.file("test-data", "Multivariate_trace", "", package = "BayesFMMM")
n_files <- 1
basis_degree <- 3
boundary_knots <- c(0, 1000)
internal_knots <- c(250, 500, 750)
X <- matrix(seq(-2, 2, 0.2), ncol = 1)

## Get CI for mean function
LL <- FLLik(dir, n_files, basis_degree, boundary_knots, internal_knots,
            time, Y, X = X, cov_adj = T)
}
