% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FSamplePaths}
\alias{FSamplePaths}
\title{Estimated Sample Paths}
\usage{
FSamplePaths(
  dir,
  n_files,
  basis_degree,
  boundary_knots,
  internal_knots,
  time,
  alpha = 0.05,
  burnin_prop = 0.1,
  simultaneous = FALSE,
  X = NULL,
  cov_adj = FALSE
)
}
\arguments{
\item{dir}{String containing the directory where the MCMC files are located}

\item{n_files}{Int containing the number of files per parameter}

\item{basis_degree}{Int containing the degree of B-splines used}

\item{boundary_knots}{Vector containing the boundary points of our index domain of interest}

\item{internal_knots}{Vector location of internal knots for B-splines}

\item{time}{Field of vectors containing time points at which the function was observed}

\item{alpha}{Double specifying the percentile of the credible interval ((1 - alpha) * 100 percent)}

\item{burnin_prop}{Double containing proportion of MCMC samples to discard}

\item{simultaneous}{Boolean indicating whether or not the credible intervals should be simultaneous credible intervals or pointwise credible intervals}

\item{X}{Matrix of covariates (each row contains the covariates for a single observation) (optional arugment)}

\item{cov_adj}{Boolean containing whether the model fit had a covariance structure that is covariate-dependent (optional argument)}
}
\value{
CPO Vector containing the CPO for each observation

//' @section Warning:
The following must be true:
\describe{
  \item{\code{n_files}}{must be an integer larger than or equal to 1}
  \item{\code{basis_degree}}{must be an integer larger than or equal to 1}
  \item{\code{internal_knots}}{must lie in the range of \code{boundary_knots}}
  \item{\code{alpha}}{must be between 0 and 1}
  \item{\code{burnin_prop}}{must be less than 1 and greater than or equal to 0}
  \item{\code{X}}{must have the same number of columns as covariates in the model (D)}
}
}
\description{
Calculates CIs for the estimated sample paths of the observed data for posterior
predictive checks.
}
\examples{
#########################
### Not Covariate Adj ###
#########################

## Set Hyperparameters
dir <- system.file("test-data", "Functional_trace", "", package = "BayesFMMM")
n_files <- 1
time <- readRDS(system.file("test-data", "time.RDS", package = "BayesFMMM"))
basis_degree <- 3
boundary_knots <- c(0, 1000)
internal_knots <- c(250, 500, 750)

## Get CI for Sample Paths
CI <- FSamplePaths(dir, n_files, basis_degree, boundary_knots, internal_knots, time)

#####################
### Covariate Adj ###
#####################

dir <- system.file("test-data", "Functional_trace", "", package = "BayesFMMM")
n_files <- 1
time <- readRDS(system.file("test-data", "time.RDS", package = "BayesFMMM"))
basis_degree <- 3
boundary_knots <- c(0, 1000)
internal_knots <- c(250, 500, 750)
X <- matrix(rnorm(40), ncol = 1)

## Get CI for Sample Paths
CI <- FSamplePaths(dir, n_files, basis_degree, boundary_knots,
                   internal_knots, time, X = X)

#####################################################################
### Covariate Adj  (with Covariate-depenent covariance structure) ###
#####################################################################

dir <- system.file("test-data", "Functional_trace", "", package = "BayesFMMM")
n_files <- 1
time <- readRDS(system.file("test-data", "time.RDS", package = "BayesFMMM"))
basis_degree <- 3
boundary_knots <- c(0, 1000)
internal_knots <- c(250, 500, 750)
X <- matrix(rnorm(40), ncol = 1)

## Get CI for Sample Paths
CI <- FSamplePaths(dir, n_files, basis_degree, boundary_knots,
                   internal_knots, time, X = X, cov_adj = TRUE)

}
