% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_ind}
\alias{get_ind}
\title{Gets index of corresponding covariance matrix}
\arguments{
\item{Z}{Vector of a row of the Z matrix}
}
\value{
ind int that can be used with the Map to find covariance matrix index
}
\description{
Gets index of corresponding covariance matrix
}
