% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getCov}
\alias{getCov}
\title{Calculates covariance matrix}
\usage{
getCov(Z, Phi, Rho, Cov)
}
\arguments{
\item{Z}{Vector that conatins the class memebership}

\item{Phi}{Cube that contains the K covariance matrices}

\item{Rho}{Matrix with each row containing the elements of the upper triangular matrix}

\item{Cov}{Matrix acting as placeholder for covariance matrix}
}
\description{
Calculates covariance matrix
}
