% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetCovCI_S}
\alias{GetCovCI_S}
\title{Calculates the Simultaneous credible interval for the covariance function between two clusters}
\usage{
GetCovCI_S(dir, n_files, n_MCMC, time1, time2, l, m)
}
\arguments{
\item{dir}{String containing the directory where the MCMC files are located}

\item{n_files}{Int containing the number of files per parameter}

\item{n_MCMC}{Int containing the number of saved MCMC iterations per file}

\item{time1}{Vector containing time points of interest for first cluster}

\item{time2}{Vector containing time points of interest for second cluster}

\item{l}{Int containing the 1st cluster group of which you want to get the credible interval for}

\item{m}{Int containing the 2nd cluster group of which you want to get the credible interval for}
}
\value{
CI list containing the 97.5th , 50th, and 2.5th simultaneous credible functions
}
\description{
Calculates the Simultaneous credible interval for the covariance function between two clusters
}
