% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetDistance}
\alias{GetDistance}
\title{Gets the distance used in ZInitialState}
\arguments{
\item{means}{Vector containing the mean of the cluster of interest}

\item{theta}{Matrix containing the weights for the smoothed observations}

\item{B_obs}{Field of Matrices containing basis functions evaluated at observed time points}

\item{dist}{Vector acting as a placeholder for the returned distances}
}
\description{
Gets the distance used in ZInitialState
}
