% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetMeanCI_Pw}
\alias{GetMeanCI_Pw}
\alias{GetMeanCI_PW}
\title{Calculates the Pointwise credible interval for the mean}
\usage{
GetMeanCI_PW(dir, n_files, time, k)
}
\arguments{
\item{dir}{String containing the directory where the MCMC files are located}

\item{n_files}{Int containing the number of files per parameter}

\item{time}{Vector containing time points of interest}

\item{k}{Int containing the cluster group of which you want to get the credible interval for}
}
\value{
CI list containing the 97.5th , 50th, and 2.5th pointwise functions
}
\description{
Calculates the Pointwise credible interval for the mean
}
