% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetStuff}
\alias{GetStuff}
\alias{TestEstimateInitialZ}
\alias{TestEstimateInitial}
\alias{TestEstimateInitialTT}
\title{Tests BFOC function}
\usage{
GetStuff()

TestEstimateInitialZ()

TestEstimateInitial(tot_mcmc_iters, r_stored_iters, directory)

TestEstimateInitialTT(tot_mcmc_iters, r_stored_iters, directory)
}
\description{
Tests BFOC function

Tests BFOC function

Tests BFOC function

Tests BFOC function
}
