% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetZCI}
\alias{GetZCI}
\title{Calculates the credible interval for membership parameters Z}
\usage{
GetZCI(dir, n_files)
}
\arguments{
\item{dir}{String containing the directory where the MCMC files are located}

\item{n_files}{Integer containing the number of files per parameter}
}
\value{
CI List containing the 97.5th , 50th, and 2.5th credible values
}
\description{
Calculates the credible interval for membership parameters Z
}
