% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lpdf_a1}
\alias{lpdf_a1}
\title{computes the log pdf of a_1j}
\arguments{
\item{alpha_1l}{Double containing hyperparameter for a}

\item{beta_1l}{Double containing hyperparameter for a}

\item{a}{Double containing value of a}

\item{delta}{double contianing value of delta}
}
\description{
computes the log pdf of a_1j
}
