% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lpdf_a2}
\alias{lpdf_a2}
\title{computes the log pdf of a_2j}
\arguments{
\item{alpha_2l}{Double containing hyperparameter for a}

\item{beta_2l}{Double containing hyperparameter for a}

\item{a}{Double containing value of a}

\item{delta}{double contianing value of delta}
}
\description{
computes the log pdf of a_2j
}
