% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lpdf_pi_PM}
\alias{lpdf_pi_PM}
\title{Calculates the log pdf of the posterior distribution of pi}
\arguments{
\item{c}{vector containing hyperparameters}

\item{alpha_3}{Double containing current value of alpha_3}

\item{pi}{Vector containing pi parameters}

\item{Z}{Matrix containing current Z parameters}
}
\value{
lpdf Double containing the log pdf
}
\description{
Calculates the log pdf of the posterior distribution of pi
}
