% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lpdf_z}
\alias{lpdf_z}
\title{Gets log-pdf of z_i given zeta_{-z_i}}
\arguments{
\item{y_obs}{Vector containing y at observed time points}

\item{y_star}{Vector containing y at unobserved time points}

\item{B_obs}{Matrix containing basis functions evaluated at observed time points}

\item{B_star}{Matrix containing basis functions evaluated at unobserved time points}

\item{Phi}{Cube containing Phi parameters}

\item{nu}{Matrix containing nu parameters}

\item{pi}{vector containing the elements of pi}

\item{Z}{Vector containing the ith row of Z}

\item{sigma_sq}{double containing the sigma_sq variable}
}
\value{
lpdf_z double contianing the log-pdf

lpdf_z double contianing the log-pdf
}
\description{
Gets log-pdf of z_i given zeta_{-z_i}

Gets log-pdf of z_i given zeta_{-z_i}
}
