% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lpdf_zMV}
\alias{lpdf_zMV}
\title{Gets log-pdf of z_i given zeta_{-z_i} for multivariate model}
\arguments{
\item{y_obs}{Vector containing ith observed point}

\item{Phi}{Cube containing Phi parameters}

\item{nu}{Matrix containing nu parameters}

\item{pi}{vector containing the elements of pi}

\item{Z}{Vector containing the ith row of Z}

\item{sigma_sq}{double containing the sigma_sq variable}
}
\value{
lpdf_z double containing the log-pdf

lpdf_z double containing the log-pdf
}
\description{
Gets log-pdf of z_i given zeta_{-z_i} for multivariate model

Gets log-pdf of z_i given zeta_{-z_i} for the multivariate model
}
