% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Model_LLik}
\alias{Model_LLik}
\title{Calculates the log-likelihood of the parameters for each iteration}
\usage{
Model_LLik(
  dir,
  n_files,
  n_MCMC,
  basis_degree,
  boundary_knots,
  internal_knots,
  time,
  Y
)
}
\arguments{
\item{dir}{String containing the directory where the MCMC files are located}

\item{n_files}{Int containing the number of files per parameter}

\item{n_MCMC}{Int containing the number of saved MCMC iterations per file}

\item{basis_degree}{Int containing the degree of B-splines used}

\item{boundary_knots}{Vector containing the boundary points of our index domain of interest}

\item{internal_knots}{Vector location of internal knots for B-splines}

\item{time}{Field of vectors containing time points at which the function was observed}

\item{Y}{Field of vectors containing observed values of the function}
}
\value{
LLik Vector containing the log-likelihood evaluated at each iteration
}
\description{
Calculates the log-likelihood of the parameters for each iteration
}
