% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MV_Model_AIC}
\alias{MV_Model_AIC}
\title{Calculates the AIC of a multivariate model}
\usage{
MV_Model_AIC(dir, n_files, n_MCMC, Y, burnin_prop = 0.2)
}
\arguments{
\item{dir}{String containing the directory where the MCMC files are located}

\item{n_files}{Int containing the number of files per parameter}

\item{n_MCMC}{Int containing the number of saved MCMC iterations per file}

\item{Y}{Matrix of observed vectors (each row is an observation)}

\item{burnin_prop}{Double containing proportion of MCMC samples to discard}
}
\value{
AIC Double containing AIC value
}
\description{
Calculates the AIC of a multivariate model
}
