% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MVCovCI}
\alias{MVCovCI}
\title{Calculates the credible interval for the covariance (Multivariate Data)}
\usage{
MVCovCI(
  dir,
  n_files,
  n_MCMC,
  l,
  m,
  alpha = 0.05,
  rescale = TRUE,
  burnin_prop = 0.1
)
}
\arguments{
\item{dir}{String containing the directory where the MCMC files are located}

\item{n_files}{Int containing the number of files per parameter}

\item{n_MCMC}{Int containing the number of saved MCMC iterations per file}

\item{l}{Int containing the 1st cluster group of which you want to get the credible interval for}

\item{m}{Int containing the 2nd cluster group of which you want to get the credible interval for}

\item{alpha}{Double specifying the percentile of the credible interval ((1 - alpha) * 100 percent)}

\item{rescale}{Boolean indicating whether or not we should rescale the Z variables so that there is at least one observation almost completely in one group}

\item{burnin_prop}{Double containing proportion of MCMC samples to discard}
}
\value{
CI list containing the credible interval for the mean function, as well as the median posterior estimate of the mean function
}
\description{
This function calculates a credible interval for the covariance matrix
between the l-th and m-th clusters, with the user specified coverage.
In order to run this function, the directory of the posterior samples needs
to be specified. The function will return the credible intervals and the median
posterior estimate of the mean. The user can specify if they would like the
algorithm to automatically rescale the parameters for interpretability
(suggested). If the user chooses to rescale, then all class memberships will
be rescaled so that at least one observation is in only one class.
}
\section{Warning}{

The following must be true:
\describe{
  \item{\code{n_files}}{must be an integer larger than or equal to 1}
  \item{\code{n_MCMC}}{must be an integer larger than or equal to 1}
  \item{\code{l}}{must be an integer larger than 1 and less than or equal to the number of clusters in the model}
  \item{\code{m}}{must be an integer larger than 1 and less than or equal to the number of clusters in the model}
  \item{\code{alpha}}{must be between 0 and 1}
  \item{\code{burnin_prop}}{must be less than 1 and greater than or equal to 0}
}
}

\examples{
## Set Hyperparameters
dir <- system.file("test-data","", package = "BayesFPMM")
n_files <- 1
n_MCMC <- 200
l <- 1
m <- 1
## Get CI for mean function
CI <- MVCovCI(dir, n_files, n_MCMC, l, m)

}
