% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MVDIC}
\alias{MVDIC}
\title{Calculates the DIC of a multivariate model}
\usage{
MVDIC(dir, n_files, Y, burnin_prop = 0.1, X = NULL, cov_adj = FALSE)
}
\arguments{
\item{dir}{String containing the directory where the MCMC files are located}

\item{n_files}{Int containing the number of files per parameter}

\item{Y}{Matrix of observed vectors (each row is an observation)}

\item{burnin_prop}{Double containing proportion of MCMC samples to discard}

\item{X}{Matrix of covariates, where each row corresponds to an observation (if covariate adjusted)}

\item{cov_adj}{Boolean containing whether or not the covariance structure depends on the covariates of interest}
}
\value{
DIC Double containing DIC value
}
\description{
Calculates the DIC of a multivariate model
}
\examples{
#########################
### Not Covariate Adj ###
#########################

## Observed data
Y <- readRDS(system.file("test-data", "MVSim_data.RDS", package = "BayesFMMM"))

## Set Hyperparameters
dir <- paste0(system.file("test-data", "Multivariate_trace", package = "BayesFMMM"), "/")
n_files <- 1

## Get DIC
DIC <- MVDIC(dir, n_files, Y)

#####################
### Covariate Adj ###
#####################

## Observed data and observed covariates
Y <- readRDS(system.file("test-data", "MVSim_data.RDS", package = "BayesFMMM"))
X <- matrix(rnorm(20, 0 , 1), nrow = 20, ncol = 1)

dir <- paste0(system.file("test-data", "Multivariate_trace", package = "BayesFMMM"), "/")
n_files <- 1

## Get DIC
DIC <- MVDIC(dir, n_files, Y, X = X)

#####################################################################
### Covariate Adj  (with Covariate-depenent covariance structure) ###
#####################################################################

## Observed data and observed covariates
Y <- readRDS(system.file("test-data", "MVSim_data.RDS", package = "BayesFMMM"))
X <- matrix(rnorm(20, 0 , 1), nrow = 20, ncol = 1)

dir <- paste0(system.file("test-data", "Multivariate_trace", package = "BayesFMMM"), "/")
n_files <- 1

## Get DIC
DIC <- MVDIC(dir, n_files, Y, X = X, cov_adj = TRUE)

}
