% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MVLLik}
\alias{MVLLik}
\title{Calculates the Log-Likelihood of a Multivariate Model}
\usage{
MVLLik(dir, n_files, Y, X = NULL, cov_adj = FALSE)
}
\arguments{
\item{dir}{String containing the directory where the MCMC files are located}

\item{n_files}{Int containing the number of files per parameter}

\item{Y}{Matrix of observed vectors (each row is an observation)}

\item{X}{Matrix of covariates (each row contains the covariates for a single observation) (optional arugment)}

\item{cov_adj}{Boolean containing whether the model fit had a covariance structure that is covariate-dependent (optional argument)}
}
\value{
LLik Vector containing the log-likelihood evaluated at each iteration
}
\description{
Calculates the log-likelihood of the parameters for each iteration of a multivariate model.
This function can handle covariate adjusted models as well as non-adjusted models.
}
\section{Warning}{

The following must be true:
\describe{
  \item{\code{n_files}}{must be an integer larger than or equal to 1}
  \item{\code{X}}{must have the same number of columns as covariates in the model (D)}
}
}

\examples{
#########################
### Not Covariate Adj ###
#########################

## Observed data
Y <- readRDS(system.file("test-data", "MVSim_data.RDS", package = "BayesFMMM"))

## Set Hyperparameters
dir <- paste0(system.file("test-data", "Multivariate_trace", package = "BayesFMMM"), "/")
n_files <- 1

## Get log likelihood
LL <- MVLLik(dir, n_files, Y)

#####################
### Covariate Adj ###
#####################

## Observed data and observed covariates
Y <- readRDS(system.file("test-data", "MVSim_data.RDS", package = "BayesFMMM"))
X <- matrix(rnorm(20, 0 , 1), nrow = 20, ncol = 1)

## Set Hyperparameters
dir <- paste0(system.file("test-data", "Multivariate_trace", package = "BayesFMMM"), "/")
n_files <- 1

## Get log likelihood
LL <- MVLLik(dir, n_files, Y, X = X)

#####################################################################
### Covariate Adj  (with Covariate-depenent covariance structure) ###
#####################################################################

## Observed data and observed covariates
Y <- readRDS(system.file("test-data", "MVSim_data.RDS", package = "BayesFMMM"))
X <- matrix(rnorm(20, 0 , 1), nrow = 20, ncol = 1)

## Set Hyperparameters
dir <- paste0(system.file("test-data", "Multivariate_trace", package = "BayesFMMM"), "/")
n_files <- 1

## Get log likelihood
LL <- MVLLik(dir, n_files, Y, X = X, cov_adj = TRUE)

}
