% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NuInitialState}
\alias{NuInitialState}
\title{Returns the estimated variance used for initial state of MCMC}
\arguments{
\item{B_obs}{Field of Matrices containing the basis functions evaluated at the observed time points}

\item{z_known}{Matrix containing the Z values for known (or estimated) functions}

\item{theta}{Matrix containing weights for basis expansion}

\item{n_funct}{Int containing the number of observed functions}
}
\value{
nu Matrix conating intial starting state for nu parameter in MCMC
}
\description{
Returns the estimated variance used for initial state of MCMC
}
