% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pi_proposal_density}
\alias{pi_proposal_density}
\title{Calculates the probability that we propose a state given we are in another state}
\usage{
pi_proposal_density(pi, alpha)
}
\arguments{
\item{pi}{Vector containing state we are proposing}

\item{alpha}{Vector containing parameters used to propose the proposed state}
}
\description{
Calculates the probability that we propose a state given we are in another state
}
