% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Generates a random sample from the Dirichlet Distribution}
\usage{
rdirichlet(alpha)
}
\arguments{
\item{alpha}{Vector containing concentration parameters}
}
\value{
distribution Vector containing the random sample
}
\description{
Generates a random sample from the Dirichlet Distribution
}
