% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReadCube}
\alias{ReadCube}
\title{Reads in armadillo cube and returns it in R format}
\usage{
ReadCube(file)
}
\arguments{
\item{file}{String containing location where arma cube is stored}
}
\description{
Reads in armadillo cube and returns it in R format
}
