% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReadField}
\alias{ReadField}
\title{Reads in armadillo field and returns it in R format}
\usage{
ReadField(file)
}
\arguments{
\item{directory}{String containing location where arma field is stored}
}
\description{
Reads in armadillo field and returns it in R format
}
