% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReadFieldCube}
\alias{ReadFieldCube}
\title{Reads in armadillo field of cubes and returns it in R format}
\usage{
ReadFieldCube(file)
}
\arguments{
\item{directory}{String containing location where arma field is stored}
}
\description{
Reads in armadillo field of cubes and returns it in R format
}
