% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReadFieldMat}
\alias{ReadFieldMat}
\title{Reads saved armadillo data}
\usage{
ReadFieldMat(file)
}
\arguments{
\item{file}{String containing location where armadillo field of matrices is stored}
}
\value{
FieldMatrix List of matrices containing the saved data
}
\description{
Reads armadillo field of matrices type data and returns it as a list of matrices
in R.
}
\examples{
## set file path
file <- system.file("test-data", "fieldmat.txt", package = "BayesFMMM")

## Read in file
samp_data <- ReadFieldMat(file)

}
