% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReadFieldVec}
\alias{ReadFieldVec}
\title{Reads in armadillo field of vectors and returns it in R format}
\usage{
ReadFieldVec(file)
}
\arguments{
\item{directory}{String containing location where arma field is stored}
}
\description{
Reads in armadillo field of vectors and returns it in R format
}
