% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReadMat}
\alias{ReadMat}
\title{Reads in armadillo matrix and returns it in R format}
\usage{
ReadMat(file)
}
\arguments{
\item{file}{String containing location where arma matrix is stored}
}
\description{
Reads in armadillo matrix and returns it in R format
}
