% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReadVec}
\alias{ReadVec}
\title{Reads in armadillo vector and returns it in R format}
\usage{
ReadVec(file)
}
\arguments{
\item{file}{String containing location where arma vector is stored}
}
\description{
Reads in armadillo vector and returns it in R format
}
