% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rmvnormal}
\alias{Rmvnormal}
\title{Generates (degenerate) multivariate normal random variable}
\usage{
Rmvnormal(mu, sigma)
}
\arguments{
\item{mu}{Vector containing mean vector}

\item{sigma}{matrix containing}

\item{n}{Int containing the number of draws}
}
\description{
Generates (degenerate) multivariate normal random variable
}
