% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SigmaCI}
\alias{SigmaCI}
\title{Calculates the credible interval for sigma squared for all types of data}
\usage{
SigmaCI(dir, n_files, alpha = 0.05)
}
\arguments{
\item{dir}{String containing the directory where the MCMC files are located}

\item{n_files}{Integer containing the number of MCMC files}

\item{alpha}{Double specifying the percentile of the credible interval ((1 - alpha) * 100 percent)}
}
\value{
CI list containing the credible interval for the covariance function, as well as the median posterior estimate of the covariance function
}
\description{
Calculates the credible interval for sigma squared for all types of data
}
\section{Warning}{

The following must be true:
\describe{
  \item{\code{n_files}}{must be an integer larger than or equal to 1}
  \item{\code{alpha}}{must be between 0 and 1}
}
}

\examples{
## Set Hyperparameters
dir <- system.file("test-data", "Multivariate_trace", "", package = "BayesFMMM")
n_files <- 1

## Get CI for Z
CI <- SigmaCI(dir, n_files)

}
