% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SigmaInitialState}
\alias{SigmaInitialState}
\title{Returns the estimated variance used for initial state of MCMC}
\arguments{
\item{y_obs}{Field of Vectors containing the observed values}

\item{B_obs}{Field of Matrices containing the basis functions evaluated at the observed time points}

\item{theta}{Matrix containing weights for basis expansion}

\item{n_funct}{Int containing the number of observed functions}
}
\value{
sigma Double containing initial starting state of sigma for MCMC
}
\description{
Returns the estimated variance used for initial state of MCMC
}
