% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TestBFPMM}
\alias{TestBFPMM}
\title{Tests the full Bayesian Functional Partial Membership Model}
\usage{
TestBFPMM(tot_mcmc_iters, r_stored_iters, directory, sigma_sq)
}
\description{
Tests the full Bayesian Functional Partial Membership Model
}
