% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TestBFPMM_MTT}
\alias{TestBFPMM_MTT}
\title{Tests mixed sampling from the Bayesian Functional Partial Membership Model}
\usage{
TestBFPMM_MTT(
  beta_N_t,
  N_t,
  n_temp_trans,
  tot_mcmc_iters,
  r_stored_iters,
  directory,
  sigma_sq
)
}
\description{
Tests mixed sampling from the Bayesian Functional Partial Membership Model
}
