% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TestBFPMM_Nu_Z}
\alias{TestBFPMM_Nu_Z}
\alias{TestBFPMM_Nu_Z_multiple_try}
\title{Tests the full Bayesian Functional Partial Membership Model}
\usage{
TestBFPMM_Nu_Z(tot_mcmc_iters, sigma_sq, beta_N_t, N_t, n_temp_trans)

TestBFPMM_Nu_Z_multiple_try(
  tot_mcmc_iters,
  sigma_sq,
  beta_N_t,
  N_t,
  n_temp_trans,
  n_trys,
  k,
  dir
)
}
\description{
Tests the full Bayesian Functional Partial Membership Model

Tests the full Bayesian Functional Partial Membership Model
}
