% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TestBFPMM_Theta}
\alias{TestBFPMM_Theta}
\title{Tests the full Bayesian Functional Partial Membership Model}
\usage{
TestBFPMM_Theta(tot_mcmc_iters, sigma_sq, Z_samp, nu_samp, burnin_prop, k)
}
\description{
Tests the full Bayesian Functional Partial Membership Model
}
