% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TestBSpline}
\alias{TestBSpline}
\alias{TestBSplineTensor}
\alias{TestPmat}
\title{Tests creation of B-splines}
\usage{
TestBSpline()

TestBSplineTensor()

TestPmat()
}
\description{
Tests creation of B-splines

Tests creation of B-splines

Tests creation of B-splines
}
