% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TestUpdateZSingleMat}
\alias{TestUpdateZSingleMat}
\title{Tests updating Z using single covariance matrix}
\usage{
TestUpdateZSingleMat()
}
\description{
Tests updating Z using single covariance matrix
}
