% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TestUpdateZSingleMatNoUnobs}
\alias{TestUpdateZSingleMatNoUnobs}
\title{Tests updating Z using single covariance matrix with no unobserved Observations}
\usage{
TestUpdateZSingleMatNoUnobs()
}
\description{
Tests updating Z using single covariance matrix with no unobserved Observations
}
