% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{UpdateA}
\alias{UpdateA}
\title{updates the a parameters for individualized covariance matrix}
\arguments{
\item{alpha_1l}{double containing hyperparameters for a}

\item{beta_1l}{double containing hyperparameters for a}

\item{alpha_2l}{double containing hyperparameters for a}

\item{beta_2l}{double containing hyperparameters for a}

\item{delta}{vec contianing value of delta}

\item{var_epsilon1}{double containing hyperparameter epsilon1}

\item{var_epsilon2}{double containing hyperparameter epsilon2}

\item{iter}{int containing MCMC iteration}

\item{a}{mat containing values of a}
}
\description{
updates the a parameters for individualized covariance matrix

updates the a parameters for single covariance matrix
}
