% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updatePi}
\alias{updatePi}
\title{Updates the parameters pi_1 through pi_K}
\usage{
updatePi(Z, alpha, K, iter, pi)
}
\arguments{
\item{Z}{Binary matrix of containing inclusion/exclusion to the various groups}

\item{alpha}{Hyperparameter for distribution of pi_i}

\item{K}{Number of clusters}

\item{iter}{Iteration of MCMC step}

\item{pi}{Matrix conatining all past, current, and future MCMC draws}
}
\description{
Updates the parameters pi_1 through pi_K
}
