% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{UpdatePi_PM}
\alias{UpdatePi_PM}
\title{Updates pi for the partial membership model}
\arguments{
\item{alpha_3}{Double containing the current value of alpha_3}

\item{Z}{Matrix containing current value of Z parameters}

\item{c}{Vector containing hyperparameters}

\item{iter}{Int containing current MCMC iteration}

\item{tot_mcmc_iters}{Int  containing total number of MCMC iterations}

\item{a_pi_PM}{Double containing hyperparameter for sampling from pi}

\item{pi_ph}{Vector containing placeholder for proposed update}

\item{pi}{Matrix containing all values for pi values}
}
\description{
Updates pi for the partial membership model
}
