% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{UpdateZMV}
\alias{UpdateZMV}
\title{Updates the Z Matrix for multivariate model}
\arguments{
\item{y_obs}{Matrix containing observed vectors}

\item{B_obs}{Field of Matrices containing basis functions evaluated at observed time points}

\item{Phi}{Cube containing Phi parameters}

\item{nu}{Matrix containing nu parameters}

\item{pi}{Vector containing the elements of pi}

\item{sigma_sq}{Double containing the sigma_sq variable}

\item{rho}{Double containing hyperparameter for proposal of new z_i state}

\item{iter}{Int containing current mcmc iteration}

\item{tot_mcmc_iters}{Int containing total number of mcmc iterations}

\item{Z_ph}{Matrix that acts as a placeholder for Z}

\item{Z}{Cube that contains all past, current, and future MCMC draws}
}
\description{
Updates the Z Matrix for multivariate model
}
