% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Z_proposal_density}
\alias{Z_proposal_density}
\title{Calculates the probability that we propose a state given we are in another state}
\usage{
Z_proposal_density(Z, alpha)
}
\arguments{
\item{Z}{Vector containing state we are proposing}

\item{alpha}{Vector containing parameters used to propose the proposed state}
}
\description{
Calculates the probability that we propose a state given we are in another state
}
