% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ZCI}
\alias{ZCI}
\title{Calculates the credible interval for membership parameters Z}
\usage{
ZCI(dir, n_files, alpha = 0.05, rescale = TRUE, burnin_prop = 0.1)
}
\arguments{
\item{dir}{String containing the directory where the MCMC files are located}

\item{n_files}{Integer containing the number of files per parameter}

\item{alpha}{Double specifying the percentile of the credible interval ((1 - alpha) * 100 percent)}

\item{rescale}{Boolean indicating whether or not we should rescale the Z variables so that there is at least one observation almost completely in one group}

\item{burnin_prop}{Double containing proportion of MCMC samples to discard}
}
\value{
CI List containing the desired credible values
}
\description{
This function constructs credible intervals using the MCMC samples of the
parameters. This function will handle high dimensional functional data,
functional data, and multivariate data.
}
