% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ZInitialState}
\alias{ZInitialState}
\title{Estimates the initial state for Z matrix in MCMC}
\arguments{
\item{B_obs}{Field of Matrices containing basis functions evaluated at observed time points}

\item{theta}{Matrix containing the weights for the smoothed observations}

\item{max_iter}{Int containing max number of iterations of algorithm}

\item{K}{Int containing number of clusters}

\item{n_funct}{Int containing the number of functions}

\item{convergence}{Double containing the criterion for convergence}
}
\value{
Z Matrix that contains the initial starting stae for Z matrix in MCMC
}
\description{
Estimates the initial state for Z matrix in MCMC
}
