\name{BayesFMMM-package}
\alias{BayesFMMM-package}
\alias{BayesFMMM}
\docType{package}
\title{
\packageTitle{BayesFMMM}
}
\description{
\packageDescription{BayesFMMM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BayesFMMM}
\packageIndices{BayesFMMM}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{BayesFMMM}

Maintainer: \packageMaintainer{BayesFMMM}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
}

