\name{BayesFPMM-package}
\alias{BayesFPMM-package}
\alias{BayesFPMM}
\docType{package}
\title{
\packageTitle{BayesFPMM}
}
\description{
\packageDescription{BayesFPMM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BayesFPMM}
\packageIndices{BayesFPMM}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{BayesFPMM}

Maintainer: \packageMaintainer{BayesFPMM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
