% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BFPMM_Nu_Z_multiple_try}
\alias{BFPMM_Nu_Z_multiple_try}
\title{Function for finding a good initial starting point for nu parameters and Z parameters, with option for temperered transitions}
\usage{
BFPMM_Nu_Z_multiple_try(
  tot_mcmc_iters,
  n_try,
  k,
  Y,
  time,
  n_funct,
  basis_degree,
  n_eigen,
  boundary_knots,
  internal_knots
)
}
\arguments{
\item{tot_mcmc_iters}{Int conatining the number of MCMC iterations per try}

\item{n_try}{Int containing how many different chains are tried}

\item{k}{Int containing the number of clusters}

\item{Y}{Field of vectors containing the observed values}

\item{time}{Field of vecotrs containing the observed time points}

\item{n_funct}{Int containing the number of functions}

\item{n_eigen}{Int containing the number of eigenfunctions}

\item{boundary_knots}{Vector containing the boundary points of our index domain of interest}

\item{internal_knots}{Vector location of internal knots for B-splines}

\item{basis}{degree Int containing the degree of B-splines used}
}
\value{
BestChain List containing a summary of the best performing chain
}
\description{
Function for finding a good initial starting point for nu parameters and Z parameters, with option for temperered transitions
}
