% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BFPMM_warm_start}
\alias{BFPMM_warm_start}
\title{Performs MCMC for functional data, with optional tempered transitions, using user specified starting points}
\usage{
BFPMM_warm_start(
  beta_N_t,
  N_t,
  n_temp_trans,
  tot_mcmc_iters,
  r_stored_iters,
  Z_samp,
  pi_samp,
  alpha_3_samp,
  delta_samp,
  gamma_samp,
  Phi_samp,
  A_samp,
  nu_samp,
  tau_samp,
  sigma_samp,
  chi_samp,
  burnin_prop,
  k,
  Y,
  time,
  n_funct,
  basis_degree,
  n_eigen,
  boundary_knots,
  internal_knots,
  thinning_num,
  dir
)
}
\arguments{
\item{beta_N_t}{Double containing the maximum weight for tempered transisitons}

\item{N_t}{Int containing total number of tempered transitions. If no tempered transitions are desired, pick a small integer}

\item{n_temp_trans}{Int containing how often tempered transitions are performed. If no tempered transitions are desired, pick a integer larger than tot_mcmc_iters}

\item{tot_mcmc_iters}{Int containing the number of MCMC iterations}

\item{r_stored_iters}{Int containing number of MCMC iterations stored in memory before writing to directory}

\item{Z_samp}{Cube containing initial chain of Z parameters}

\item{pi_samp}{Matrix containing initial chain of pi parameters}

\item{alpha_3_samp}{Vector containing initial chain of alpha_3 parameters}

\item{delta_samp}{Matrix containing initial chain of delta parameters}

\item{gamma_samp}{Field of cubes containing initial chain of gamma parameters}

\item{Phi_samp}{Field of cubes containing initial chain of phi parameters}

\item{A_samp}{Matrix containing initial chain of A parameters}

\item{nu_samp}{Cube containing initial chain of nu parameters}

\item{tau_samp}{Matrix containing initial chain of tau parameters}

\item{sigma_samp}{Vector containing initial chain of sigma parameters}

\item{chi_samp}{Cube containing initial chain of chi parameters}

\item{burnin_prop}{Double containing proportion of chain used to estimate the starting point of nu parameters and Z parameters}

\item{k}{Int containing the number of clusters}

\item{Y}{Field of vectors containing the observed values}

\item{time}{Field of vectors containing the observed time points}

\item{n_funct}{Int containing the number of functions}

\item{basis_degree}{Int containing the degree of B-splines used}

\item{n_eigen}{Int containing the number of eigenfunctions}

\item{boundary_knots}{Vector containing the boundary points of our index domain of interest}

\item{internal_knots}{Vector location of internal knots for B-splines}

\item{thinning_num}{Int containing how often we should save MCMC iterations. Should be a divisible by r_stored_iters and tot_mcmc_iters}

\item{dir}{String containing directory where the MCMC files should be saved}
}
\description{
Performs MCMC for functional data, with optional tempered transitions, using user specified starting points
}
