% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_m}
\alias{compute_m}
\title{Computes all m vectors}
\usage{
compute_m(S_obs, Z, phi, nu, m)
}
\arguments{
\item{S_obs}{Field of Matrices containing basis functions evaluated at observed time points}

\item{Z}{Matrix of current Z parameter}

\item{phi}{Cube of current phi paramaters}

\item{nu}{Matrix of current nu parameters}

\item{m}{matrix acting as a placeholder for m}
}
\description{
Computes all m vectors
}
