% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{g_ldet}
\alias{g_ldet}
\title{Gets generalized log determinant (product of positive eigen values)}
\usage{
g_ldet(S)
}
\arguments{
\item{S}{Vector of singular values}
}
\value{
g_ldet Double countaining the generalized determinant
}
\description{
Gets generalized log determinant (product of positive eigen values)
}
