% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lpdf_z}
\alias{lpdf_z}
\title{Gets log-pdf of z given zeta}
\usage{
lpdf_z(M, m, tilde_M, tilde_m, f_obs, f_star, pi_l, z_il, pinv_M, pinv_tilde_M)
}
\arguments{
\item{M}{Cube that contains the M_i variance matrices}

\item{m}{Matrix that contains the m_i mean vectors}

\item{tilde_M}{Cube that contains the tilde_M_i variance matrices}

\item{tilde_m}{Matrix that contains the tilde_M_i mean vectors}

\item{f_obs}{Vector containing the observed points for the ith observation}

\item{f_star}{Vector containing the unobserved time points of interest for ith observation}

\item{pi}{Vector containing the sampled pi for this iteration}
}
\value{
lpdf_z Double contianing the log-pdf
}
\description{
Gets log-pdf of z given zeta
}
