% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lpdf_zTemperedMV}
\alias{lpdf_zTemperedMV}
\title{Gets log-pdf of z_i given zeta_{-z_i} using tempered transitions for the multivariate model}
\arguments{
\item{beta_i}{Double containing current temperature}

\item{y_obs}{Vector containing the ith observed point}

\item{Phi}{Cube containing Phi parameters}

\item{nu}{Matrix containing nu parameters}

\item{pi}{vector containing the elements of pi}

\item{Z}{Vector containing the ith row of Z}

\item{sigma_sq}{double containing the sigma_sq variable}
}
\value{
lpdf_z double containing the log-pdf

lpdf_z double containing the log-pdf
}
\description{
Gets log-pdf of z_i given zeta_{-z_i} using tempered transitions for the multivariate model

Gets log-pdf of z_i given zeta_{-z_i} using tempered transitions for the multivariate model
}
