% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Model_DIC}
\alias{Model_DIC}
\alias{MV_Model_DIC}
\title{Calculates the DIC of a functional model}
\usage{
MV_Model_DIC(dir, n_files, n_MCMC, Y, burnin_prop = 0.2)
}
\arguments{
\item{dir}{String containing the directory where the MCMC files are located}

\item{n_files}{Int containing the number of files per parameter}

\item{n_MCMC}{Int containing the number of saved MCMC iterations per file}

\item{Y}{Matrix of observed vectors (each row is an observation)}

\item{burnin_prop}{Double containing proportion of MCMC samples to discard}
}
\value{
DIC Double containing DIC value
}
\description{
Calculates the DIC of a functional model
}
