% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Model_DIC}
\alias{Model_DIC}
\title{Calculates the DIC of a model}
\usage{
Model_DIC(dir, n_files, n_MCMC, time, Y)
}
\arguments{
\item{dir}{String containing the directory where the MCMC files are located}

\item{n_files}{Int containing the number of files per parameter}

\item{n_MCMC}{Int containing the number of saved MCMC iterations per file}

\item{time}{Field of vectors containing time points at which the function was observed}

\item{Y}{Field of vectors containing observed values of the function}
}
\value{
DIC Double containing DIC value
}
\description{
Calculates the DIC of a model
}
