% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MV_Model_LLik}
\alias{MV_Model_LLik}
\title{Calculates the log-likelihood of the parameters for each iteration of a multivariate model}
\usage{
MV_Model_LLik(dir, n_files, n_MCMC, Y)
}
\arguments{
\item{dir}{String containing the directory where the MCMC files are located}

\item{n_files}{Int containing the number of files per parameter}

\item{n_MCMC}{Int containing the number of saved MCMC iterations per file}

\item{Y}{Matrix of observed vectors (each row is an observation)}
}
\value{
LLik Vector containing the log-likelihood evaluated at each iteration
}
\description{
Calculates the log-likelihood of the parameters for each iteration of a multivariate model
}
