% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MVCovCI}
\alias{MVCovCI}
\title{Calculates the credible interval for the covariance (Multivariate Data)}
\usage{
MVCovCI(
  dir,
  n_files,
  l,
  m,
  alpha = 0.05,
  rescale = TRUE,
  burnin_prop = 0.1,
  X = NULL
)
}
\arguments{
\item{dir}{String containing the directory where the MCMC files are located}

\item{n_files}{Int containing the number of files per parameter}

\item{l}{Int containing the 1st cluster group of which you want to get the credible interval for}

\item{m}{Int containing the 2nd cluster group of which you want to get the credible interval for}

\item{alpha}{Double specifying the percentile of the credible interval ((1 - alpha) * 100 percent)}

\item{rescale}{Boolean indicating whether or not we should rescale the Z variables so that there is at least one observation almost completely in one group}

\item{burnin_prop}{Double containing proportion of MCMC samples to discard}

\item{X}{Matrix containing covariates at points of interest (of dimension W x D (number of points of interest x number of covariates))}
}
\value{
CI list containing the credible interval for the mean function, as well as the median posterior estimate of the mean function. Posterior estimates of the covariance function are also returned.
}
\description{
This function calculates a credible interval for the covariance matrix
between the l-th and m-th clusters, with the user specified coverage. This
function can handle covariate adjusted models, where the mean and covariance
functions depend on the covariates of interest. If not covariate adjusted, or
if the covariates only influence the mean structure, DO NOT specify \code{X} in
this function.
In order to run this function, the directory of the posterior samples needs
to be specified. The function will return the credible intervals and the median
posterior estimate of the mean. The user can specify if they would like the
algorithm to automatically rescale the parameters for interpretability
(suggested). If the user chooses to rescale, then all class memberships will
be rescaled so that at least one observation is in only one class.
}
\section{Warning}{

The following must be true:
\describe{
  \item{\code{n_files}}{must be an integer larger than or equal to 1}
  \item{\code{n_MCMC}}{must be an integer larger than or equal to 1}
  \item{\code{l}}{must be an integer larger than 1 and less than or equal to the number of clusters in the model}
  \item{\code{m}}{must be an integer larger than 1 and less than or equal to the number of clusters in the model}
  \item{\code{alpha}}{must be between 0 and 1}
  \item{\code{burnin_prop}}{must be less than 1 and greater than or equal to 0}
  \item{\code{X}}{must have the same number of columns as covariates in the model (D)}
}
}

\examples{
#########################
### Not Covariate Adj ###
#########################

## Set Hyperparameters
dir <- paste0(system.file("test-data", "Multivariate_trace", package = "BayesFMMM"), "/")
n_files <- 1
l <- 1
m <- 1

## Get CI for cov function
CI <- MVCovCI(dir, n_files, l, m)

#####################
### Covariate Adj ###
#####################

dir <- paste0(system.file("test-data", "Multivariate_trace", package = "BayesFMMM"), "/")
n_files <- 1
l <- 1
m <- 1

## Get CI for cov function
CI <- MVCovCI(dir, n_files, l, m)

#####################################################################
### Covariate Adj  (with Covariate-depenent covariance structure) ###
#####################################################################

dir <- paste0(system.file("test-data", "Multivariate_trace", package = "BayesFMMM"), "/")
n_files <- 1
l <- 1
m <- 1
X <- matrix(seq(-2, 2, 0.2), ncol = 1)

## Get CI for cov function
CI <- MVCovCI(dir, n_files, l, m, X = X)

}
