% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MVMeanCI}
\alias{MVMeanCI}
\title{Calculates the credible interval for the mean (Multivariate Data)}
\usage{
MVMeanCI(dir, n_files, alpha = 0.05, rescale = TRUE, burnin_prop = 0.1)
}
\arguments{
\item{dir}{String containing the directory where the MCMC files are located}

\item{n_files}{Int containing the number of files per parameter}

\item{alpha}{Double specifying the percentile of the credible interval ((1 - alpha) * 100 percent)}

\item{rescale}{Boolean indicating whether or not we should rescale the Z variables so that there is at least one observation almost completely in one group}

\item{burnin_prop}{Double containing proportion of MCMC samples to get rid}
}
\value{
CI list containing the credible interval for the mean function, as well as the median posterior estimate of the mean function
}
\description{
This function calculates a credible interval with the user specified coverage.
In order to run this function, the directory of the posterior samples needs
to be specified. The function will return the credible intervals and the median
posterior estimate of the mean. The user can specify if they would like the
algorithm to automatically rescale the parameters for interpretability
(suggested). If the user chooses to rescale, then all class memberships will
be rescaled so that at least one observation is in only one class.
}
\section{Warning}{

The following must be true:
\describe{
  \item{\code{n_files}}{must be an integer larger than or equal to 1}
  \item{\code{alpha}}{must be between 0 and 1}
  \item{\code{burnin_prop}}{must be less than 1 and greater than or equal to 0}
}
}

\examples{
## Set Hyperparameters
dir <- system.file("test-data","", package = "BayesFPMM")
n_files <- 1

## Get CI for mean function
CI <- MVMeanCI(dir, n_files)

}
