% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReadFieldMat}
\alias{ReadFieldMat}
\title{Reads in armadillo field of matrices and returns it in R format}
\usage{
ReadFieldMat(file)
}
\arguments{
\item{directory}{String containing location where arma field is stored}
}
\description{
Reads in armadillo field of matrices and returns it in R format
}
