% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rmvnormal}
\alias{Rmvnormal}
\title{Generates multivariate normal random variable}
\arguments{
\item{mu}{Vector containing mean vector}

\item{sigma}{Matrix containing covariance matrix}
}
\description{
Generates multivariate normal random variable
}
