% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rmvnormal}
\alias{Rmvnormal}
\title{Generates (degenerate) multivariate normal random variable}
\usage{
Rmvnormal(mu, sigma)
}
\arguments{
\item{mu}{Vector containing mean vector}

\item{sigma}{matrix containing covariance matrix}

\item{U}{Matrix acting as a placeholder for SVD}

\item{S}{Vector acting as a placeholder for SVD}

\item{V}{Matrix acting as a placeholder for SVD}
}
\description{
Generates (degenerate) multivariate normal random variable

Generates (degenerate) multivariate normal random variable
}
