% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TestBFPMM_MTT}
\alias{TestBFPMM_MTT}
\alias{TestBFPMM_MTT_warm_start}
\title{Tests mixed sampling from the Bayesian Functional Partial Membership Model}
\usage{
TestBFPMM_MTT(
  beta_N_t,
  N_t,
  n_temp_trans,
  tot_mcmc_iters,
  r_stored_iters,
  directory,
  sigma_sq
)

TestBFPMM_MTT_warm_start(
  beta_N_t,
  N_t,
  n_temp_trans,
  tot_mcmc_iters,
  r_stored_iters,
  directory,
  sigma_sq,
  Z_samp,
  pi_samp,
  alpha_3_samp,
  delta_samp,
  gamma_samp,
  Phi_samp,
  A_samp,
  nu_samp,
  tau_samp,
  sigma_samp,
  chi_samp,
  burnin_prop,
  k,
  dir
)
}
\description{
Tests mixed sampling from the Bayesian Functional Partial Membership Model

Tests mixed sampling from the Bayesian Functional Partial Membership Model
}
