% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updateA}
\alias{updateA}
\title{updates the a parameters for individualized covariance matrix}
\arguments{
\item{alpha_1l}{Double containing hyperparameters for a}

\item{beta_1l}{Double containing hyperparameters for a}

\item{alpha_2l}{Double containing hyperparameters for a}

\item{beta_2l}{Double containing hyperparameters for a}

\item{delta}{Vector contianing value of delta}

\item{var_epsilon1}{Double containing hyperparameter epsilon1}

\item{var_epsilon2}{Double containing hyperparameter epsilon2}

\item{iter}{Double containing MCMC iteration}

\item{a}{Mat containing values of a}
}
\description{
updates the a parameters for individualized covariance matrix
}
