% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{UpdatePi}
\alias{UpdatePi}
\title{Updates pi}
\arguments{
\item{alpha}{Double that is the hyperparameter}

\item{iter}{Int containing current MCMC iteration}

\item{tot_mcmc_iters}{Int  containing total number of MCMC iterations}

\item{Z}{Matrix that contains the current values of the binary matrix}

\item{pi}{Matrix containg all samples of Pi}
}
\description{
Updates pi
}
