% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updatePi}
\alias{updatePi}
\title{Updates the parameters pi_1 through pi_K}
\usage{
updatePi(Z, alpha, K, N, iter, pi)
}
\arguments{
\item{Z}{Binary matrix of containing inclusion/exclusion to the various groups}

\item{alpha}{Hyperparameter for distribution of pi_i}

\item{K}{Number of clusters}

\item{pi_l}{Placeholder of new draw}
}
\value{
updated pi_l
}
\description{
Updates the parameters pi_1 through pi_K
}
