% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{UpdateZ}
\alias{UpdateZ}
\title{Updates the Z Matrix}
\arguments{
\item{f_obs}{Field of vectors containing f at observed time points}

\item{pi}{Vector containing the sampled pi for this iteration}

\item{iter}{Iteration of MCMC step}

\item{S_obs}{Field of Matrices containing basis functions evaluated at observed time points}

\item{Phi}{Cube of current Phi paramaters}

\item{Rho}{Matrix with each row containing the elements of the upper triangular matrix}

\item{nu}{Matrix that contains all current nu paramaters}

\item{Cov}{Matrix containing placeholder for covariance matrix}

\item{m}{Field of Vectors that contains all m mean vectors}

\item{mean_ph_obs}{Field of vectors that serve as a placeholder of computations}

\item{Z_ph}{Matrix that acts as a placeholder for Z}

\item{Z}{Cube that contains all past, current, and future MCMC draws}
}
\description{
Updates the Z Matrix
}
