% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{UpdateZ}
\alias{UpdateZ}
\title{Updates the Z Matrix}
\arguments{
\item{y_obs}{Field of Vectors containing y at observed time points}

\item{y_star}{Field of Matrices containing y at unobserved time points at all mcmc iterations}

\item{B_obs}{Field of Matrices containing basis functions evaluated at observed time points}

\item{B_star}{Field of Matrices containing basis functions evaluated at unobserved time points}

\item{Phi}{Cube containing Phi parameters}

\item{nu}{Matrix containing nu parameters}

\item{pi}{Vector containing the elements of pi}

\item{sigma_sq}{Double containing the sigma_sq variable}

\item{rho}{Double containing hyperparameter for proposal of new z_i state}

\item{n_known}{Int containing number of observations with known membership}

\item{iter}{Int containing current mcmc iteration}

\item{tot_mcmc_iters}{Int containing total number of mcmc iterations}

\item{Z_ph}{Matrix that acts as a placeholder for Z}

\item{Z}{Cube that contains all past, current, and future MCMC draws}
}
\description{
Updates the Z Matrix

Updates the Z Matrix when we have a subset of known class memberships
}
