% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updateZ}
\alias{updateZ}
\title{Updates the Z Matrix}
\usage{
updateZ(
  f_obs,
  f_star,
  pi,
  iter,
  S_obs,
  S_star,
  phi,
  nu,
  M,
  M_ph,
  pinv_M,
  m,
  m_ph,
  Z_ph,
  tilde_M,
  tilde_M_ph,
  pinv_tilde_M,
  tilde_m,
  tilde_m_ph,
  mp_inv,
  Z
)
}
\arguments{
\item{f_obs}{Field of vectors containing f at observed time points}

\item{f_star}{Field of vectors containing f at unobserved time points}

\item{pi}{Vector containing the sampled pi for this iteration}

\item{iter}{Iteration of MCMC step}

\item{S_obs}{Field of Matrices containing basis functions evaluated at observed time points}

\item{S_star}{Field of Matrices containing basis functions evaluated at unobserved time points}

\item{phi}{Cube of current phi paramaters}

\item{nu}{Matrix that contains all current nu paramaters}

\item{M}{Field of Matrices that contains all M matrices}

\item{M_ph}{Field of Matrices that acts as a placeholder for the new M matrix}

\item{pinv_M}{Field of Matrices that acts as a placeholder for g-inverse of M}

\item{m}{Field of Vectors that contains all m mean vectors}

\item{m_ph}{Field of Vectors that acts as a placeholder for new m vector}

\item{Z_ph}{Matrix that acts as a placeholder for the new Z matrix}

\item{tilde_M}{Field of Matrices that contains the tilde_M_i variance matrix}

\item{tilde_M_ph}{Field of Matrices that acts as a placeholder for tilde_M}

\item{pinv_tilde_M}{Field of Matrices that acts as a placeholder for g-inverse of tilde M}

\item{tilde_m}{Vector that contains the tilde_M_i mean vector}

\item{Z}{Cube that contains all past, current, and future MCMC draws}

\item{z_ph}{Matrix that acts as a placeholder for Z}

\item{Z_plus}{Field of Matrices acting as a placeholder for Z_plus}

\item{A_plus}{Matrix acting as a placeholder for A_plus}

\item{C}{Matrix acting as a placeholder for C}
}
\description{
Updates the Z Matrix
}
