% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updateZ_i}
\alias{updateZ_i}
\title{Updates the ith row of the Z Matrix}
\usage{
updateZ_i(M, m, tilde_M, tilde_m, f_obs, f_star, pi, iter, Z)
}
\arguments{
\item{M}{Matrix that contains the M_i variance matrix}

\item{m}{Vector that contains the m_i mean vector}

\item{tilde_M}{Matrix that contains the tilde_M_i variance matrix}

\item{tilde_m}{Vector that contains the tilde_M_i mean vector}

\item{f_obs}{Vector containing the observed points n}

\item{f_star}{Vector containing the unobserved time points of interest for ith observation}

\item{pi}{Vector containing the sampled pi for this iteration}

\item{iter}{Iteration of MCMC step}

\item{Z}{Cube that contains all past, current, and future MCMC draws}
}
\description{
Updates the ith row of the Z Matrix
}
